### 2024-07-15 #6977 @MicahGale
> MISSING DOIs
>
> - No DOI given, and none found for title: MCNP User’s Manual: Code Version 6.2

The MCNP 6.3 manual does have a DOI number on OSTI. I will have to check but changing to citing the 6.3 manual will probably make sense.

> - No DOI given, and none found for title: SLY (Sly Lex Yacc)

Nothing can be done here.

> - No DOI given, and none found for title: Advanced Test Reactor User Guide

This report is on [OSTI](https://www.osti.gov/biblio/1826354) as an INL external report, and it's odd that it doesn't have a DOI. I will ask around to see if this can be given a DOI.
