### 2024-07-15 #6977 @MicahGale
> Hello @MicahGale, sorry for the delay in processing this. Unfortunately, we do not have any editors available to handle your submission right now, so I have to put this on our waitlist. In the meantime, any reviewer recommendations would be welcome (though please don't actually tag/request folks yourself).

Thank you @kyleniemeyer for the update, and moving this along for the time being. I will get you a list of potential reviewers in the next few days.

Just as a heads up I will be on vacation starting next week through August 12, and won't be checking GitHub during that time. If needed @tjlaboss should be able to address most concerns about this paper during that time.
