### 2024-07-17 #6977 @tjlaboss
> The MCNP 6.3 manual does have a DOI number on OSTI. I will have to check but changing to citing the 6.3 manual will probably make sense.

This makes sense to me. Not only is MCNP6.3 the most recent version at time of submission, but also its documentation has been cleaned up and expanded since 6.2.
