### 2024-07-20 #6977 @cmarmo
Hello @kyleniemeyer , Chiara speaking here, I am currently serving as editor in chief for pyOpenSci: @MicahGale submitted MontePy to pyOpenSci, without knowing we have a collaboration with JOSS.

I cannot guarantee to find an editor faster than JOSS, but moving forward via pyOpenSci we can probably optimize the number of volounteers mobilized for this review: would that sound reasonable?
