### 2024-08-12 #6977 @kyleniemeyer
@cmarmo sorry, I just saw your note about pyOpenSci. Normally, for rOpenSci and pyOpenSci, our fast-track process kicks in after the package was reviewed and approved there, to avoid reviewing the same things twice. Then, the JOSS editor essentially just checks for scope and then reviews the paper.

If @MicahGale is fine with it, we can pause the JOSS review until the pyOpenSci review has concluded, and then handle the paper fast-track acceptance afterwards. (This will probably take the same amount of time anyways!)
