### 2024-08-12 #6977 @cmarmo
@kyleniemeyer , I'm having trouble finding editors this summer, but I think the "pyOpenSci then JOSS" workflow can save reviewers some time. On the other hand, you've already found an editor.

So, I dare to imagine a scenario... let me know if it could be feasible...

What if @kellyrowland would be interested in being an editor for the pyOpenSci submission?
It looks to me like she would be a perfect fit! And we are always happy to have new collaborators!
