### 2024-11-18 #6977 @kellyrowland
link to pyOpenSci issue: https://github.com/pyOpenSci/software-submission/issues/205
