### 2025-01-05 16b8b5f
Updated changelog to RTD links.
