+++ doc/source/changelog.rst
@@ -37,7 +37,7 @@ MontePy Changelog
 **Performance Improvement**
 
 * Fixed cyclic memory reference that lead to memory leak in ``copy.deepcopy`` (:issue:`514`).
-* Fixed O(N<sup>2</sup>) operation in how append works for object collections like Cells (:issue:`556`).
+* Fixed O(N\ :sup:`2`) operation in how append works for object collections like Cells (:issue:`556`).
 
 **Bug Fixes**
 
