### 2025-01-11 7f18ee7
Added append option to MCNP_Problem.parse.
