### 2025-01-11 6b9baf7
Promoted parsing functions to be top level functions.
