+++ montepy/__init__.py
@@ -18,6 +18,7 @@ from montepy.data_inputs.material import Material
 from montepy.data_inputs.transform import Transform
 from montepy.data_inputs.nuclide import Library, Nuclide
 from montepy.data_inputs.element import Element
+from montepy.data_inputs.thermal_scattering import ThermalScatteringLaw
 
 # geometry
 from montepy.geometry_operators import Operator
