### 2025-01-11 565d19c
Made parse_surface function name more pythonic.
