### 2025-01-11 843780d
Added demo of append option.
