### 2025-01-11 8b184de
Updated tests for pretty_str change.
