+++ tests/test_syntax_parsing.py
@@ -440,7 +440,7 @@ class TestSyntaxNode(TestCase):
     def test_syntax_str(self):
         str(self.test_node)
         repr(self.test_node)
-        self.test_node.pretty_str()
+        self.test_node._pretty_str()
 
 
 class TestGeometryTree(TestCase):
@@ -479,7 +479,7 @@ class TestGeometryTree(TestCase):
         test = self.test_tree
         str(test)
         repr(test)
-        test.pretty_str()
+        test._pretty_str()
 
     def test_geometry_comments(self):
         test = copy.deepcopy(self.test_tree)
@@ -698,7 +698,7 @@ class TestListNode(TestCase):
         list_node.append(syntax_node.ValueNode("1.0", float))
         str(list_node)
         repr(list_node)
-        list_node.pretty_str()
+        list_node._pretty_str()
 
     def test_list_slicing(self):
         list_node = syntax_node.ListNode("list")
@@ -796,7 +796,7 @@ class TestMaterialssNode(TestCase):
         isotopes.append_nuclide(("isotope_fraction", zaid, concentration))
         str(isotopes)
         repr(isotopes)
-        isotopes.pretty_str()
+        isotopes._pretty_str()
 
     def test_isotopes_iter(self):
         isotopes = syntax_node.MaterialsNode("test")
@@ -1577,7 +1577,7 @@ class TestParametersNode(TestCase):
     def test_parameter_str(self):
         str(self.param)
         repr(self.param)
-        self.param.pretty_str()
+        self.param._pretty_str()
 
     def test_parameter_format(self):
         self.assertEqual(self.param.format(), "vol=1.0")
