### 2025-01-11 c8abb71
Updated references to deprecated surface_builder.
