### 2025-01-11 4838a86
Removed all version change markers for 0.2.0
I figure that 0.2.0 was really the second major release,
so marking the changes from the first to second major release now that
we are on the third is no longer necessary.
