+++ montepy/data_inputs/data_parser.py
@@ -32,9 +32,6 @@ def parse_data(input: montepy.mcnp_object.InitInput):
     """
     Parses the data input as the appropriate object if it is supported.
 
-    .. versionchanged:: 0.2.0
-        Removed the ``comment`` parameter, as it's in the syntax tree directly now.
-
     :param input: the Input object for this Data input
     :type input: Union[Input, str]
     :return: the parsed DataInput object
