+++ montepy/mcnp_object.py
@@ -68,7 +68,6 @@ class _ExceptionContextAdder(ABCMeta):
         """
         This will replace all properties and callable attributes with
         wrapped versions.
-
         """
         new_attrs = {}
         for key, value in attributes.items():
@@ -97,10 +96,6 @@ class _ExceptionContextAdder(ABCMeta):
 class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
     """
     Abstract class for semantic representations of MCNP inputs.
-
-    .. versionchanged:: 0.2.0
-        Generally significant changes for parser rework.
-        For init removed ``comments``, and added ``parser`` as arguments.
 
     :param input: The Input syntax object this will wrap and parse.
     :type input: Union[Input, str]
@@ -177,8 +172,6 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
 
         None is generally a safe default value to provide.
 
-        .. versionadded:: 0.2.0
-
         :param value_type: the data type for the ValueNode.
         :type value_type: Class
         :param default: the default value to provide (type needs to agree with value_type)
@@ -219,8 +212,6 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
         The most common need is to update a value based on the number for an object pointed at,
         e.g., the material number in a cell definition.
 
-        .. versionadded:: 0.2.0
-
         """
         pass
 
@@ -257,8 +248,6 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
     def leading_comments(self) -> list[PaddingNode]:
         """
         Any comments that come before the beginning of the input proper.
-
-        .. versionadded:: 0.2.0
 
         :returns: the leading comments.
         :rtype: list
