+++ montepy/input_parser/syntax_node.py
@@ -25,9 +25,6 @@ class SyntaxNodeBase(ABC):
     A syntax node is any component of the syntax tree
     for a parsed input.
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :param name: a name for labeling this node.
     :type name: str
     """
@@ -225,9 +222,6 @@ class SyntaxNode(SyntaxNodeBase):
         if key in syntax_node:
             pass
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :param name: a name for labeling this node.
     :type name: str
     :param parse_dict: the dictionary of the syntax tree nodes.
@@ -340,9 +334,6 @@ class SyntaxNode(SyntaxNodeBase):
 class GeometryTree(SyntaxNodeBase):
     """
     A syntax tree that is a binary tree for representing CSG geometry logic.
-
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
 
     .. versionchanged:: 0.4.1
         Added left/right_short_type
@@ -601,9 +592,6 @@ class PaddingNode(SyntaxNodeBase):
     """
     A syntax tree node to represent a collection of sequential padding elements.
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :param token: The first padding token for this node.
     :type token: str
     :param is_comment: If the token provided is a comment.
@@ -789,9 +777,6 @@ class CommentNode(SyntaxNodeBase):
     """
     Object to represent a comment in an MCNP problem.
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :param input: the token from the lexer
     :type input: Token
     """
@@ -908,9 +893,6 @@ class ValueNode(SyntaxNodeBase):
 
     This stores the original input token, the current value,
     and the possible associated padding.
-
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
 
     :param token: the original token for the ValueNode.
     :type token: str
@@ -1418,9 +1400,6 @@ class ParticleNode(SyntaxNodeBase):
     """
     A node to hold particles information in a :class:`ClassifierNode`.
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     :param name: the name for the node.
     :type name: str
     :param token: the original token from parsing
@@ -1553,9 +1532,6 @@ class ParticleNode(SyntaxNodeBase):
 class ListNode(SyntaxNodeBase):
     """
     A node to represent a list of values.
-
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
 
     :param name: the name of this node.
     :type name: str
@@ -1901,9 +1877,6 @@ class ShortcutNode(ListNode):
     A node that pretends to be a :class:`ListNode` but is actually representing a shortcut.
 
     This takes the shortcut tokens, and expands it into their "virtual" values.
-
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
 
     :param p: the parsing object to parse.
     :type p: sly.yacc.YaccProduction
@@ -2361,9 +2334,6 @@ class ClassifierNode(SyntaxNodeBase):
     """
     A node to represent the classifier for a :class:`montepy.data_input.DataInput`
 
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
-
     e.g., represents ``M4``, ``F104:n,p``, ``IMP:n,e``.
     """
 
@@ -2530,9 +2500,6 @@ class ParametersNode(SyntaxNodeBase):
     A node to hold the parameters, key-value pairs, for this input.
 
     This behaves like a dictionary and is accessible by their key*
-
-    .. versionadded:: 0.2.0
-        This was added with the major parser rework.
 
     .. Note::
         How to access values.
