### 2025-01-22 96c3405
Removed unneeded settings.
