### 2025-01-22 c12bcdc
Updated to latest sigstore due to bug.
