+++ .github/workflows/deploy-alpha.yml
@@ -1,4 +1,4 @@
-name: Deploy
+name: Alpha-Deploy
 
 on: 
   push:
@@ -43,7 +43,7 @@ jobs:
         run: .github/scripts/check_version.py --alpha 
       - run: python -m build .
       - name: Sign the dists with Sigstore
-        uses: sigstore/gh-action-sigstore-python@v2.1.1
+        uses: sigstore/gh-action-sigstore-python@v3.0.0
         with:
           inputs: >-
             ./dist/*.tar.gz
