### 2025-01-25 211221e
Made it possible to have a blank sdef.
