### 2025-01-29 640db53
Added SSR keywords and made formatting consistant.
