### 2025-01-29 1ce8ccb
Documented optional dependencies.
