+++ inst/paper.bib
@@ -81,6 +81,7 @@
    title = {The MCNPTools Package: Installation and Use},
    institution = {Los Alamos National Laboratory},
    DOI = {10.2172/1884737},
+   number = {LA-UR-22-28935},
    url = {https://www.osti.gov/biblio/1884737},
    year = {2022},
    type = {Report}
