### 2025-01-29 c967b8e
Added bib info for PyNE.
