+++ inst/paper.bib
@@ -101,3 +101,13 @@
 }
 
 
+@misc{Scopatz:2012,
+   author = {Scopatz, Anthony and Romano, Paul K. and Wilson, Paul P. H. and Huff, Kathryn D.},
+   title = {PyNE: Python for Nuclear Engineering},
+   publisher = {American Nuclear Society},
+   volume = {107},
+   pages = {985-987},
+   month = {November 11-15},
+   year = {2012},
+   type = {Conference Paper}
+}
