### 2025-01-30 6b1ac88
Added strict parameter.
