### 2025-02-03 5f367c0
added citations for all the competitors I could find.
