+++ inst/paper.bib
@@ -111,3 +111,601 @@
    year = {2012},
    type = {Conference Paper}
 }
+
+@software{rodionov_mckit_2024,
+	title = {mckit},
+	url = {https://github.com/MC-kit/mckit},
+	version = {0.8.3},
+	author = {Rodionov, Roman and Portnov, Dmitri},
+	date = {2024},
+}
+
+@software{travleev_numjuggler_nodate,
+	title = {numjuggler},
+	url = {https://github.com/travleev/numjuggler},
+	version = {2.42.36},
+	author = {Travleev, Anton and Previti, Alberto and Portnov, Dmitri},
+}
+
+@software{mariano_mcnp_2022,
+	title = {{MCNP} Input Reader},
+	url = {https://github.com/ENEA-Fusion-Neutronics/MCNP-Input-Reader},
+	version = {0.2.1},
+	author = {Mariano, Giovanni},
+	date = {2022},
+	keywords = {read input},
+}
+
+@software{laghi_mctools_2023,
+	title = {mctools},
+	url = {https://github.com/dodu94/mctools},
+	author = {Laghi, Davide},
+	date = {2023},
+}
+
+@software{batkov_mc-tools_2024,
+	title = {mc-tools},
+	url = {https://github.com/kbat/mc-tools},
+	version = {1.0.post1},
+	author = {Batkov, Konstantin and Borghi, Nicolò and Furutaka, Kazuyoshi and Ansell, Stuart and Vezhlev, Egor},
+	date = {2024-12-20},
+}
+
+@software{pacific_northwest_national_laboratory_cardsharpformcnp_2025,
+	title = {{CardSharpForMCNP}},
+	url = {https://github.com/pnnl/CardSharpForMCNP},
+	version = {1.4.2},
+	author = {{Pacific Northwest National Laboratory}},
+	date = {2025},
+}
+
+@software{hagen_wig_2021,
+	title = {wig},
+	url = {https://github.com/alexhagen/wig},
+	author = {Hagen, Alex},
+	date = {2021-04},
+}
+
+@inproceedings{persaud_python-based_2024,
+	title = {Python-based software tools for {MCNP}},
+	url = {https://doi.org/10.1109/NSS/MIC/RTSD57108.2024.10655039},
+	doi = {10.1109/NSS/MIC/RTSD57108.2024.10655039},
+	pages = {1--1},
+	booktitle = {2024 {IEEE} Nuclear Science Symposium ({NSS}), Medical Imaging Conference ({MIC}) and Room Temperature Semiconductor Detector Conference ({RTSD})},
+	author = {Persaud, A. and Unzueta, M. Ayllon and Surry, E. K. and Parsons, A. M.},
+	date = {2024},
+	keywords = {Data visualization, Feature extraction, Geometry, Microwave integrated circuits, Monte Carlo methods, Semiconductor detectors, Writing},
+}
+
+@software{marcinkevicius_npp_mcnp_plugin_2025,
+	title = {{NPP}\_MCNP\_Plugin},
+	url = {https://github.com/kordusas/npp_mcnp_plugin},
+	author = {Marcinkevicius, Benjaminas},
+	date = {2025},
+	keywords = {syntax highlighter},
+}
+
+@article{harb_novel_2023,
+	title = {A Novel Algorithm for {CAD} to {CSG} Conversion in {McCAD}},
+	volume = {4},
+	rights = {https://creativecommons.org/licenses/by/4.0/},
+	issn = {2673-4362},
+	url = {https://www.mdpi.com/2673-4362/4/2/31},
+	doi = {10.3390/jne4020031},
[…]
