+++ inst/paper.bib
@@ -120,10 +120,11 @@
 	date = {2024},
 }
 
-@software{travleev_numjuggler_nodate,
+@software{travleev_numjuggler_2022,
 	title = {numjuggler},
 	url = {https://github.com/travleev/numjuggler},
 	version = {2.42.36},
+	date = {2022},
 	author = {Travleev, Anton and Previti, Alberto and Portnov, Dmitri},
 }
 
