### 2025-02-03 2d7d4a8
Fixed formatting and citations.
