### 2025-02-03 47d7971
Added info about PyNE and mckit.
