### 2025-02-04 #6977 @kyleniemeyer
@MicahGale normally, we wait until the end of the reviewing process to set the version, since it's only needed before accepting (and may change following reviewer feedback)
