### 2025-02-07 72958e5
New ParticleTypeWarning as parent class
