### 2025-02-07 228272f
Treat ParticleType exceptions as warnings
