### 2025-02-07 a159e8c
Convert test_importance.py from unittest to pytest
Also, throughout, 'assertEqual' and 'assertAlmostEqual' were used
interchangeably to test floating-point importances.
I have changed them all to be 'assert x == y'.
Alternatively, one could compare values with 'pytest.approx'
