+++ tests/test_importance.py
@@ -1,5 +1,4 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
-from unittest import TestCase
 import montepy
 from montepy.cell import Cell
 from montepy.particle import Particle
@@ -7,178 +6,176 @@ from montepy.data_inputs.importance import Importance
 from montepy.errors import *
 from montepy.input_parser import mcnp_input, block_type
 import os
+import pytest
 
 
-class TestImportance(TestCase):
-    def test_importance_init_cell(self):
-        # test_normal cell init
-        in_str = "1 0 -1 IMP:N,P=1"
-        card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+def test_importance_init_cell():
+    # test_normal cell init
+    in_str = "1 0 -1 IMP:N,P=1"
+    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+    cell = Cell(card)
+    assert cell.importance.neutron == 1.0
+    assert cell.importance.photon == 1.0
+    assert cell.importance.alpha_particle == 0.0
+    assert cell.importance.all is None
+    assert cell.importance.in_cell_block
+    # test non-number imp
+    in_str = "1 0 -1 IMP:N,P=h"
+    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+    with pytest.raises(ValueError):
         cell = Cell(card)
-        self.assertEqual(cell.importance.neutron, 1.0)
-        self.assertEqual(cell.importance.photon, 1.0)
-        self.assertEqual(cell.importance.alpha_particle, 0.0)
-        self.assertIsNone(cell.importance.all)
-        self.assertTrue(cell.importance.in_cell_block)
-        # test non-number imp
-        in_str = "1 0 -1 IMP:N,P=h"
-        card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
-        with self.assertRaises(ValueError):
-            cell = Cell(card)
-        # test negative imp
-        in_str = "1 0 -1 IMP:N,P=-2"
-        card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
-        with self.assertRaises(ValueError):
-            cell = Cell(card)
+    # test negative imp
+    in_str = "1 0 -1 IMP:N,P=-2"
+    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+    with pytest.raises(ValueError):
+        cell = Cell(card)
 
-    def test_importance_init_data(self):
-        in_str = "IMP:N,P 1 0"
-        card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+
+def test_importance_init_data():
+    in_str = "IMP:N,P 1 0"
+    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+    imp = Importance(card)
+    assert [
+        val.value for val in imp._particle_importances[Particle.NEUTRON]["data"]
+    ] == [1.0, 0.0]
+    assert [
+        val.value for val in imp._particle_importances[Particle.PHOTON]["data"]
+    ] == [1.0, 0.0]
+    # test non-number imp
+    in_str = "IMP:N,P 1 h"
+    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
+    with pytest.raises(ValueError):
         imp = Importance(card)
-        self.assertEqual(
-            [val.value for val in imp._particle_importances[Particle.NEUTRON]["data"]],
-            [1.0, 0.0],
-        )
-        self.assertEqual(
-            [val.value for val in imp._particle_importances[Particle.PHOTON]["data"]],
-            [1.0, 0.0],
-        )
-        # test non-number imp
-        in_str = "IMP:N,P 1 h"
[…]
