### 2025-02-07 63ab5ee
Unit test for warning, not error
