### 2025-02-07 b343454
Make default importance a NotImplementedError
