### 2025-02-07 bfe7e94
Fixed typo in sphinx syntax.
