### 2025-02-07 e90c5a3
Test the new NotImplementedError
