+++ montepy/input_parser/input_syntax_reader.py
@@ -1,17 +1,21 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
-from .block_type import BlockType
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from collections import deque
-from .. import errors
 import itertools
 import io
+import re
+import os
+import warnings
+
+
+from .block_type import BlockType
+from .. import errors
 from montepy.constants import *
 from montepy.errors import *
 from montepy.input_parser.input_file import MCNP_InputFile
 from montepy.input_parser.mcnp_input import Input, Message, ReadInput, Title
 from montepy.input_parser.read_parser import ReadParser
 from montepy.utilities import is_comment
-import os
-import warnings
+
 
 reading_queue = []
 
@@ -179,10 +183,16 @@ def read_data(fh, mcnp_version, block_type=None, recursion=False):
         old_line = line
         line = line[:line_length]
         if len(old_line) != len(line):
-            warnings.warn(
-                f"The line: {old_line} exceeded the allowed line length of: {line_length} for MCNP {mcnp_version}",
-                errors.LineOverRunWarning,
-            )
+            if len(line.split("$")[0]) >= line_length and not COMMENT_FINDER.match(
+                line
+            ):
+                warnings.warn(
+                    f"The line: {old_line} exceeded the allowed line length of: {line_length} for MCNP {mcnp_version}",
+                    errors.LineOverRunWarning,
+                )
+            # if extra length is a comment keep it long
+            else:
+                line = old_line
         if line.endswith(" &\n"):
             continue_input = True
         else:
