### 2025-02-08 bf92f6b
Update PR template to match pinned black version
