+++ doc/source/starting.rst
@@ -617,19 +617,33 @@ Order of precedence and grouping is automatically handled by Python so you can e
 
 .. testcode::
 
+   import montepy.surfaces as surfs
+   import montepy.surfaces.surface_type import SurfaceType
+
    # build blank surfaces 
    bottom_plane = montepy.surfaces.axis_plane.AxisPlane()
+   bottom_plane.surface_type = SurfaceType.PZ
    bottom_plane.location = 0.0
+
    top_plane = montepy.surfaces.axis_plane.AxisPlane()
+   top_plane.surface_type = SurfaceType.PZ
    top_plane.location = 10.0
+
    fuel_cylinder = montepy.surfaces.cylinder_on_axis.CylinderOnAxis()
+   fuel_cylinder.surface_type = SurfaceType.CZ
    fuel_cylinder.radius = 1.26 / 2
+
    clad_cylinder = montepy.surfaces.cylinder_on_axis.CylinderOnAxis()
    clad_cylinder.radius = (1.26 / 2) + 1e-3 # fuel, gap, cladding
+   clad_cylinder.surface_type = SurfaceType.CZ
+
    clad_od = montepy.surfaces.cylinder_on_axis.CylinderOnAxis()
    clad_od.radius = clad_cylinder.radius + 0.1 # add thickness
+   clad_od.surface_type = SurfaceType.CZ
    other_fuel = montepy.surfaces.cylinder_on_axis.CylinderOnAxis()
    other_fuel.radius = 3.0
+   other_fuel_region.surface_type = SurfaceType.CZ
+
    bottom_plane.number = 1
    top_plane.number = 2
    fuel_cylinder.number = 3
