### 2025-02-09 acb31ae
Added guide to setting boundary conditions.
