### 2025-02-09 329937b
Fixed bugs in getting started.
