### 2025-02-13 d3ffb01
Added test for strict material contains.
