### 2025-02-14 2dc6a57
Added strict argument to find.
