### 2025-02-14 3758aca
Update material find tests and did type testing.
