### 2025-02-14 a294df3
Added type enforcement to strict.
