### 2025-02-14 3d9c54d
Updated docstrings for find and contains.
