### 2025-02-21 18d2d51
Tried to replicate #666
