### 2025-02-22 af15449
Added exact replication of #666 to testing.
