### 2025-02-22 44ff8c8
Implemented material clear.
