### 2025-02-22 4123432
Avoided line expansion warnings for manually set default libraries.
