### 2025-02-25 e623f4c
Switched to nuclides pattern, and made type enforcement top priority.
