### 2025-02-25 e6863ce
Propagated strict up to materials, and clarified strict.
