### 2025-02-25 308f21c
Prototyped contains all, contains any
