### 2025-02-25 b11a631
Propogated to materials for contains_any contains_all
