+++ montepy/materials.py
@@ -1,6 +1,7 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 
 from __future__ import annotations
+import collections as co
 import copy
 from typing import Generator, Union
 
@@ -30,7 +31,7 @@ class Materials(NumberedDataObjectCollection):
     def __init__(self, objects=None, problem=None):
         super().__init__(Material, objects, problem)
 
-    def get_containing(
+    def get_containing_any(
         self,
         *nuclides: Union[
             montepy.data_inputs.nuclide.Nuclide,
@@ -43,7 +44,7 @@ class Materials(NumberedDataObjectCollection):
         strict: bool = False,
     ) -> Generator[Material]:
         """
-        Get all materials that contain these nuclides.
+        Get all materials that contain any of these these nuclides.
 
         This uses :func:`~montepy.data_inputs.material.Material.contains` under the hood.
         See that documentation for more guidance.
@@ -57,7 +58,7 @@ class Materials(NumberedDataObjectCollection):
 
             import montepy
             problem = montepy.read_input("foo.imcnp")
-            for mat in problem.materials.get_containing("H-1", "O-16", threshold = 0.3):
+            for mat in problem.materials.get_containing_any("H-1", "U-235", threshold = 0.3):
                 print(mat)
 
         .. testoutput::
@@ -66,7 +67,7 @@ class Materials(NumberedDataObjectCollection):
 
         .. versionadded:: 1.0.0
 
-        :param nuclides: a plurality of other nuclides to check for.
+        :param nuclides: a plurality of nuclides to check for.
         :type nuclides: Union[Nuclide, Nucleus, Element, str, int]
         :param threshold: the minimum concentration of a nuclide to be considered. The material components are not
             first normalized.
@@ -81,27 +82,81 @@ class Materials(NumberedDataObjectCollection):
         :raises TypeError: if any argument is of the wrong type.
         :raises ValueError: if the fraction is not positive or zero, or if nuclide cannot be interpreted as a Nuclide.
         """
-        # TODO collision here
-        # TODO other type enforcement
-        nuclides = []
+        return self._contains_arb(
+            *nuclides, bool_func=any, threshold=threshold, strict=strict
+        )
+
+    def get_containing_all(
+        self,
+        *nuclides: Union[
+            montepy.data_inputs.nuclide.Nuclide,
+            montepy.data_inputs.nuclide.Nucleus,
+            montepy.Element,
+            str,
+            int,
+        ],
+        threshold: float = 0.0,
+        strict: bool = False,
+    ) -> Generator[Material]:
+        """
+        Get all materials that contain all of these nuclides.
+
+        This uses :func:`~montepy.data_inputs.material.Material.contains` under the hood.
+        See that documentation for more guidance.
+
+        Examples
+        ^^^^^^^^
+
+        One example would to be find all water bearing materials:
+
+        .. testcode::
+
+            import montepy
+            problem = montepy.read_input("foo.imcnp")
+            for mat in problem.materials.get_containing_all("H-1", "O-16", threshold = 0.3):
+                print(mat)
+
+        .. testoutput::
+
+            MATERIAL: 1, ['hydrogen', 'oxygen']
+
+        .. versionadded:: 1.0.0
+
+        :param nuclides: a plurality of nuclides to check for.
+        :type nuclides: Union[Nuclide, Nucleus, Element, str, int]
+        :param threshold: the minimum concentration of a nuclide to be considered. The material components are not
+            first normalized.
+        :type threshold: float
+        :param strict: If True this does not let an elemental nuclide match all child isotopes, isomers, nor will an isotope
+            match all isomers, nor will a blank library match all libraries.
+        :type strict: bool
+
[…]
