### 2025-02-26 12e7851
Made strict parameter more explicit and clear.
