### 2025-02-27 52ab1ae
made parsing rule for making even pair number lists.
