### 2025-02-28 6f4a70b
Added anti-particle mapping; I don't know particle physics 🙃
