### 2025-02-28 df3d0a5
Updating changelog to next version: 0.5.4
