### 2025-03-01 301c837
Fixed edge case in ParticleNode
