+++ montepy/input_parser/syntax_node.py
@@ -1483,8 +1483,9 @@ class ParticleNode(SyntaxNodeBase):
                 if match.group(0).isupper():
                     upper_match += 1
                 total_match += 1
-        if upper_match / total_match >= 0.5:
-            self._formatter["upper"] = True
+        if total_match:
+            if upper_match / total_match >= 0.5:
+                self._formatter["upper"] = True
 
     @property
     def comments(self):
