### 2025-03-03 4b1383a
Simplified nested logic.
