### 2025-03-04 ebaa972
Made sure that when adding things to a material the syntax tree is valid.
