### 2025-03-04 ed48881
Made default libraries str prettier.
