### 2025-03-05 6af8271
Added test for #667 to suppress noisy LineExpansion
