### 2025-03-05 0faceca
Converted test_syntax_parsing fully to pytest.
