+++ tests/test_syntax_parsing.py
@@ -2,7 +2,6 @@
 import copy
 from io import StringIO
 import pytest
-from unittest import TestCase
 
 import montepy
 from montepy.errors import *
@@ -17,7 +16,7 @@ from montepy.particle import Particle
 import warnings
 
 
-class TestValueNode(TestCase):
+class TestValueNode:
     def test_valuenoode_init(self):
         for type, token, answer in [
             (str, "hi", "hi"),
@@ -28,35 +27,35 @@ class TestValueNode(TestCase):
         ]:
             for padding in [None, syntax_node.PaddingNode(" ")]:
                 node = syntax_node.ValueNode(token, type, padding)
-                self.assertEqual(node.value, answer)
-                self.assertEqual(node.token, token)
+                assert node.value == answer
+                assert node.token == token
                 if padding:
-                    self.assertEqual(node.padding, padding)
+                    assert node.padding == padding
                 else:
-                    self.assertIsNone(node.padding)
+                    assert node.padding is None
         # test with None values
         for type in {str, float, int}:
             node = syntax_node.ValueNode(None, type)
-            self.assertIsNone(node.value)
+            assert None is (node.value)
             node = syntax_node.ValueNode(Jump(), type)
-            self.assertIsNone(node.value)
+            assert None is (node.value)
 
     def test_valuenode_convert_to_int(self):
         node = syntax_node.ValueNode("1", float)
         node._convert_to_int()
-        self.assertEqual(node.type, int)
-        self.assertEqual(node.value, 1)
+        assert node.type == int
+        assert node.value == 1
         # test 1.0
         node = syntax_node.ValueNode("1.0", float)
         node._convert_to_int()
-        self.assertEqual(node.type, int)
-        self.assertEqual(node.value, 1)
+        assert node.type == int
+        assert node.value == 1
         # test wrong type
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             node = syntax_node.ValueNode("hi", str)
             node._convert_to_int()
         # test real float
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             node = syntax_node.ValueNode("1.23", float)
             node._convert_to_int()
 
@@ -64,79 +63,79 @@ class TestValueNode(TestCase):
         node = syntax_node.ValueNode("1", float)
         lat = montepy.data_inputs.lattice.Lattice
         node._convert_to_enum(lat)
-        self.assertEqual(node.type, lat)
-        self.assertEqual(node.value, lat(1))
+        assert node.type == lat
+        assert node.value == lat(1)
         # test with None
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             node = syntax_node.ValueNode(None, float)
             node._convert_to_enum(lat)
         node._convert_to_enum(lat, allow_none=True)
-        self.assertIsNone(node.value)
+        assert None is (node.value)
         st = montepy.surfaces.surface_type.SurfaceType
         node = syntax_node.ValueNode("p", str)
         node._convert_to_enum(st, switch_to_upper=True)
-        self.assertEqual(node.type, st)
-        self.assertEqual(node.value, st("P"))
+        assert node.type == st
+        assert node.value == st("P")
 
     def test_is_negat_identifier(self):
         node = syntax_node.ValueNode("-1", float)
-        self.assertTrue(not node.is_negatable_identifier)
-        self.assertIsNone(node.is_negative)
+        assert not node.is_negatable_identifier
+        assert None is (node.is_negative)
         node.is_negatable_identifier = True
-        self.assertTrue(node.is_negatable_identifier)
-        self.assertEqual(node.type, int)
-        self.assertTrue(node.value > 0)
-        self.assertTrue(node.is_negative)
[…]
