### 2025-03-06 19ee2c4
Parametrized tests for easier debugging.
