### 2025-03-06 555f9c5
Clarified docstrings for @tjlaboss
