### 2025-03-06 c98dd4c
Added type Alias to make typing more DRY.
