### 2025-03-06 08e77a3
Updated tests to allow ints in contains.
