### 2025-03-07 5cd55d8
Made cumulative threshold for contains.
