### 2025-03-07 3c72ca6
Added explenation of how contains_all and elements work.
