### 2025-03-07 fc48210
Allowed mat contains to accept an int.
