+++ montepy/data_inputs/material.py
@@ -37,7 +37,7 @@ This is used for adding new material components.
 By default all components made from scratch are added to their own line with this many leading spaces.
 """
 
-NuclideLike = Union[Nuclide, Nucleus, Element, str, int]
+NuclideLike = Union[Nuclide, Nucleus, Element, str, numbers.Integral]
 
 
 class _DefaultLibraries:
@@ -610,7 +610,7 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
             raise TypeError(
                 f"Can only check if a Nuclide, Nucleus, Element, or str is in a material. {nuclide} given."
             )
-        if isinstance(nuclide, str):
+        if isinstance(nuclide, (str, numbers.Integral)):
             nuclide = Nuclide(nuclide)
         # switch to elemental
         if isinstance(nuclide, (Nucleus, Nuclide)) and nuclide.A == 0:
@@ -725,9 +725,9 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
         .. note::
 
             The difference between :func:`contains_all` and :func:`contains_any` is only for how they
-            handle being given multiple nuclides. This does not impact how given Elements will match 
+            handle being given multiple nuclides. This does not impact how given Elements will match
             daughter Nuclides. This is handled instead by ``strict``.
-        
+
         .. note::
 
             For details on how to use the ``strict`` argument see the examples in: :func:`find`.
