### 2025-03-07 0d7abdb
Removed type aliases in isinstance for py 3.9
