### 2025-03-07 f37d75f
Update surface.py isinstance
Also clean up imports and update copyright
