### 2025-03-07 ee3b0e7
Update nuclide.py isinstance
Also clean up imports, update copyright, etc.
