### 2025-03-07 c6b9b54
Update syntax_node.py isinstance
Also clean up imports, update copyright
