### 2025-03-07 79c1f88
Update cell.py isinstance
Also clean up imports, update copyright
