### 2025-03-07 1d17948
Update cells.py isinstance
Also clean up imports, update copyright
