### 2025-03-07 605b86a
Update element.py isinstance
Also update copyright
