### 2025-03-07 637b96d
Update material.py isinstance
Also clean up imports, update copyright
