+++ montepy/data_inputs/material.py
@@ -1,23 +1,18 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from __future__ import annotations
 import collections as co
 import copy
-import itertools
 import math
-import numbers
-import re
+from numbers import Integral, Real
 from typing import Generator, Union
-import warnings
 import weakref
 
 import montepy
 from montepy.data_inputs import data_input, thermal_scattering
 from montepy.data_inputs.nuclide import Library, Nucleus, Nuclide
 from montepy.data_inputs.element import Element
-from montepy.data_inputs.material_component import MaterialComponent
 from montepy.input_parser import syntax_node
 from montepy.input_parser.material_parser import MaterialParser
-from montepy import mcnp_object
 from montepy.numbered_mcnp_object import Numbered_MCNP_Object, InitInput
 from montepy.errors import *
 from montepy.utilities import *
@@ -37,7 +32,7 @@ This is used for adding new material components.
 By default all components made from scratch are added to their own line with this many leading spaces.
 """
 
-NuclideLike = Union[Nuclide, Nucleus, Element, str, numbers.Integral]
+NuclideLike = Union[Nuclide, Nucleus, Element, str, Integral]
 
 
 class _DefaultLibraries:
@@ -507,9 +502,9 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
 
     def __getitem__(self, idx):
         """ """
-        if not isinstance(idx, (int, slice)):
+        if not isinstance(idx, (Integral, slice)):
             raise TypeError(f"Not a valid index. {idx} given.")
-        if isinstance(idx, int):
+        if isinstance(idx, Integral):
             comp = self._components[idx]
             return self.__unwrap_comp(comp)
         # else it's a slice
@@ -528,7 +523,7 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
 
     def __setitem__(self, idx, newvalue):
         """ """
-        if not isinstance(idx, (int, slice)):
+        if not isinstance(idx, (Integral, slice)):
             raise TypeError(f"Not a valid index. {idx} given.")
         old_vals = self._components[idx]
         self._check_valid_comp(newvalue)
@@ -541,7 +536,7 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
     def __len__(self):
         return len(self._components)
 
-    def _check_valid_comp(self, newvalue: tuple[Nuclide, float]):
+    def _check_valid_comp(self, newvalue: tuple[Nuclide, Real]):
         """
         Checks valid compositions and raises an error if needed.
         """
@@ -555,7 +550,7 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
             )
         if not isinstance(newvalue[0], Nuclide):
             raise TypeError(f"First element must be an Nuclide. {newvalue[0]} given.")
-        if not isinstance(newvalue[1], (float, int)):
+        if not isinstance(newvalue[1], Real):
             raise TypeError(
                 f"Second element must be a fraction greater than 0. {newvalue[1]} given."
             )
@@ -565,9 +560,9 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
             )
 
     def __delitem__(self, idx):
-        if not isinstance(idx, (int, slice)):
+        if not isinstance(idx, (Integral, slice)):
             raise TypeError(f"Not a valid index. {idx} given.")
-        if isinstance(idx, int):
+        if isinstance(idx, Integral):
             self.__delitem(idx)
             return
         # else it's a slice
[…]
