### 2025-03-07 228bda1
Update numbered_mcnp_object.py isinstance
Also clean up imports, update copyright
