### 2025-03-07 89d7e49
Update numbered_object_collection.py isinstance
Also clean up imports, update copyright
