### 2025-03-07 bd057d1
Update half_space.py isinstance
Also update copyright
