+++ montepy/surfaces/half_space.py
@@ -1,4 +1,4 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from __future__ import annotations
 import montepy
 from montepy.errors import *
@@ -10,6 +10,8 @@ from montepy.input_parser.syntax_node import (
     CommentNode,
 )
 from montepy.utilities import *
+
+from numbers import Integral
 
 
 class HalfSpace:
@@ -578,7 +580,7 @@ class UnitHalfSpace(HalfSpace):
                 side = "+"
         else:
             side = "-"
-        if isinstance(self.divider, int):
+        if isinstance(self.divider, Integral):
             div = self.divider
         else:
             div = self.divider.number
@@ -643,7 +645,7 @@ class UnitHalfSpace(HalfSpace):
         if self._is_cell:
             container = cells
             par_container = self._cell.complements
-        if isinstance(self.divider, int):
+        if isinstance(self.divider, Integral):
             try:
                 self._divider = container[self._divider]
                 if self._divider not in par_container:
@@ -659,7 +661,7 @@ class UnitHalfSpace(HalfSpace):
 
     def _ensure_has_nodes(self):
         if self.node is None:
-            if isinstance(self.divider, int):
+            if isinstance(self.divider, Integral):
                 num = self.divider
             else:
                 num = self.divider.number
@@ -670,7 +672,7 @@ class UnitHalfSpace(HalfSpace):
             self._node = node
 
     def _update_node(self):
-        if isinstance(self.divider, int):
+        if isinstance(self.divider, Integral):
             self._node.value = self.divider
         else:
             self._node.value = self.divider.number
@@ -721,7 +723,7 @@ class UnitHalfSpace(HalfSpace):
         """
 
         def num(obj):
-            if isinstance(obj, int):
+            if isinstance(obj, Integral):
                 return obj
             return obj.number
 
