### 2025-03-07 f13d982
Update universe.py isinstance
Also update copyright
