### 2025-03-07 704b700
Update CHANGELOG for #679
