### 2025-03-08 81a4055
Add failing test case for issue #549 (empty Importance object print)
