### 2025-03-08 617659e
Fix issue #549: added default representation of Importance object not connected to _problem
