+++ montepy/data_inputs/importance.py
@@ -182,10 +182,31 @@ class Importance(CellModifierInput):
             raise TypeError("Key must be a particle")
         del self._particle_importances[particle]
 
+    def _format_default(self):
+        # Create a simple, self-contained string representation of the importance settings.
+        ret = []
+        for particle, tree in self._particle_importances.items():
+            # Instead of tree["classifier"].particles.value (which doesn't exist),
+            # use str(tree["classifier"].particles) or an appropriate attribute.
+            ret.append(
+                f"{tree['classifier'].prefix.value}"
+                f"{str(tree['classifier'].particles)}"
+                f"{tree['seperator'].value}"
+                f"{tree['data'].nodes[0].value}"
+            )
+        return ret
+    
     def __str__(self):
-        if not self.in_cell_block and self._problem is None:
-            return " ".join(self.input_lines)
-        return "".join(self.format_for_mcnp_input(DEFAULT_VERSION))
+        if self._problem:
+            # Use the full formatting (which relies on _problem.cells, etc.)
+            return "".join(self.format_for_mcnp_input(DEFAULT_VERSION))
+        else:
+            # Fall back to a default simple representation 
+            default = self._format_default()
+            if default:
+                return " ".join(default)
+            return "Importance object is empty"
+
 
     def __repr__(self):
         return (
