### 2025-03-08 d7aed68
Respond to @MicahGale review on copyright
and re-do changelog nubmer
