+++ montepy/surfaces/surface.py
@@ -1,4 +1,4 @@
-# Copyright 2025, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from __future__ import annotations
 import copy
 from typing import Union
