+++ montepy/data_inputs/element.py
@@ -1,4 +1,4 @@
-# Copyright 2025, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from __future__ import annotations
 from montepy.errors import *
 from montepy._singleton import SingletonGroup
