+++ montepy/input_parser/syntax_node.py
@@ -1,4 +1,4 @@
-# Copyright 2025, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from abc import ABC, abstractmethod
 import re
 import warnings
