### 2025-03-08 dccf223
improved docstring, using tree.format() to show the tree structure returing list added type hints to the function signature
