### 2025-03-10 34c8e66
Tried converting to numpy docstrings using docconvert.
