+++ montepy/input_parser/parser_base.py
@@ -8,9 +8,7 @@ _dec = sly.yacc._decorator
 
 
 class MetaBuilder(sly.yacc.ParserMeta):
-    """
-    Custom MetaClass for allowing subclassing of MCNP_Parser.
-
+    """Custom MetaClass for allowing subclassing of MCNP_Parser.
 
     Note: overloading functions is not allowed.
     """
@@ -52,9 +50,7 @@ class MetaBuilder(sly.yacc.ParserMeta):
 
 
 class SLY_Supressor:
-    """
-    This is a fake logger meant to mostly make warnings dissapear.
-    """
+    """This is a fake logger meant to mostly make warnings dissapear."""
 
     def __init__(self):
         self._parse_fail_queue = []
@@ -71,28 +67,31 @@ class SLY_Supressor:
     critical = debug
 
     def parse_error(self, msg, token=None, lineno=0, index=0):
-        """
-        Adds a SLY parsing error to the error queue for being dumped later.
+        """Adds a SLY parsing error to the error queue for being dumped later.
 
-        :param msg: The message to display.
-        :type msg: str
-        :param token: the token that caused the error if any.
-        :type token: Token
-        :param lineno: the current lineno of the error (from SLY not the file), if any.
-        :type lineno: int
+        Parameters
+        ----------
+        msg : str
+            The message to display.
+        token : Token
+            the token that caused the error if any.
+        lineno : int
+            the current lineno of the error (from SLY not the file), if
+            any.
         """
         self._parse_fail_queue.append(
             {"message": msg, "token": token, "line": lineno, "index": index}
         )
 
     def clear_queue(self):
-        """
-        Clears the error queue and returns all errors.
+        """Clears the error queue and returns all errors.
 
         Returns a list of dictionaries. The dictionary has the keys: "message", "token", "line.
 
-        :returns: A list of the errors since the queue was last cleared.
-        :rtype: list
+        Returns
+        -------
+        list
+            A list of the errors since the queue was last cleared.
         """
         ret = self._parse_fail_queue
         self._parse_fail_queue = []
@@ -103,9 +102,7 @@ class SLY_Supressor:
 
 
 class MCNP_Parser(Parser, metaclass=MetaBuilder):
-    """
-    Base class for all MCNP parsers that provides basics.
-    """
+    """Base class for all MCNP parsers that provides basics."""
 
     # Remove this if trying to see issues with parser
     log = SLY_Supressor()
@@ -113,8 +110,7 @@ class MCNP_Parser(Parser, metaclass=MetaBuilder):
     debugfile = None
 
     def restart(self):
-        """
-        Clears internal state information about the current parse.
+        """Clears internal state information about the current parse.
 
         Should be ran before a new object is parsed.
         """
@@ -122,17 +118,21 @@ class MCNP_Parser(Parser, metaclass=MetaBuilder):
         super().restart()
 
     def parse(self, token_generator, input=None):
-        """
-        Parses the token stream and returns a syntax tree.
+        """Parses the token stream and returns a syntax tree.
 
         If the parsing fails None will be returned.
         The error queue can be retrieved from ``parser.log.clear_queue()``.
 
-        :param token_generator: the token generator from ``lexer.tokenize``.
-        :type token_generator: generator
-        :param input: the input that is being lexed and parsed.
-        :type input: Input
-        :rtype: SyntaxNode
+        Parameters
+        ----------
+        token_generator : generator
+            the token generator from ``lexer.tokenize``.
+        input : Input
+            the input that is being lexed and parsed.
+
+        Returns
[…]
