+++ montepy/data_inputs/lattice.py
@@ -4,17 +4,12 @@ from enum import Enum, unique
 
 @unique
 class Lattice(Enum):
-    """
-    Represents the options for the lattice ``LAT``.
-    """
+    """Represents the options for the lattice ``LAT``."""
 
     HEXAHEDRA = 1
-    """
-    Hexhedra are solids with six faces.
+    """Hexhedra are solids with six faces.
 
     One such solid is a rectangular prism.
     """
     HEXAGONAL = 2
-    """
-    Hexagonal prism are solids with eight faces.
-    """
+    """Hexagonal prism are solids with eight faces."""

[…]
