### 2025-03-10 0a4cb4d
Formatted with black, and manually fixed a few stragglers.
