### 2025-03-10 6a25859
Added numpydoc as a dependency.
