### 2025-03-10 114aec4
Revert "Added numpydoc as a dependency."
This reverts commit 6a25859f8b5e6650eb4bd3a15b7012a76744edbe.
