+++ doc/source/conf.py
@@ -37,7 +37,6 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.extlinks",
     "sphinx.ext.doctest",
-    "numpydoc",
     "sphinx_autodoc_typehints",
     "sphinx_favicon",
     "sphinx_copybutton",
