### 2025-03-11 3e5a982
Made deprecation clear from one-line summary.
