### 2025-03-11 197913b
Don't require sphinx docs checkvalue for now
