### 2025-03-11 7a3ed5f
Fixed accidental recursion.
