### 2025-03-11 f6f9476
Rename checkvalue -> _check_value
