### 2025-03-11 f955e41
Vendor "checkvalue.py" from OpenMC and add license (#688)
* Vendor "checkvalue.py" from OpenMC and add license

* The OpenMC codebase is released under the MIT license.
This commit includes a module from OpenMC to add value checking
capabilities to MontePy. The LICENSE is updated accordingly.

* Update CHANGELOG

* Don't require sphinx docs checkvalue for now

* Rename checkvalue -> _check_value

* Update LICENSE

* Copy license VERBATIM

and add serial commas to MontePy's license

* Update LICENSE to reflect checkvalue's 2014 date
