+++ montepy/_check_value.py
@@ -1,0 +1,432 @@
+"""
+Copyright (c) 2014-2025 Massachusetts Institute of Technology, UChicago Argonne
+LLC, Battelle Energy Alliance, and OpenMC contributors
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of
+this software and associated documentation files (the "Software"), to deal in
+the Software without restriction, including without limitation the rights to
+use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+the Software, and to permit persons to whom the Software is furnished to do so,
+subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+"""
+
+import copy
+import os
+from collections.abc import Iterable
+
+import numpy as np
+
+# Type for arguments that accept file paths
+PathLike = str | os.PathLike
+
+
+def check_type(name, value, expected_type, expected_iter_type=None, *, none_ok=False):
+    """Ensure that an object is of an expected type. Optionally, if the object is
+    iterable, check that each element is of a particular type.
+
+    Parameters
+    ----------
+    name : str
+        Description of value being checked
+    value : object
+        Object to check type of
+    expected_type : type or Iterable of type
+        type to check object against
+    expected_iter_type : type or Iterable of type or None, optional
+        Expected type of each element in value, assuming it is iterable. If
+        None, no check will be performed.
+    none_ok : bool, optional
+        Whether None is allowed as a value
+
+    """
+    if none_ok and value is None:
+        return
+
+    if not isinstance(value, expected_type):
+        if isinstance(expected_type, Iterable):
+            msg = (
+                'Unable to set "{}" to "{}" which is not one of the '
+                'following types: "{}"'.format(
+                    name, value, ", ".join([t.__name__ for t in expected_type])
+                )
+            )
+        else:
+            msg = (
+                f'Unable to set "{name}" to "{value}" which is not of type "'
+                f'{expected_type.__name__}"'
+            )
+        raise TypeError(msg)
+
+    if expected_iter_type:
+        if isinstance(value, np.ndarray):
+            if not issubclass(value.dtype.type, expected_iter_type):
+                msg = (
+                    f'Unable to set "{name}" to "{value}" since each item '
+                    f'must be of type "{expected_iter_type.__name__}"'
+                )
+                raise TypeError(msg)
+            else:
+                return
+
+        for item in value:
+            if not isinstance(item, expected_iter_type):
+                if isinstance(expected_iter_type, Iterable):
+                    msg = (
+                        'Unable to set "{}" to "{}" since each item must be '
+                        'one of the following types: "{}"'.format(
+                            name,
+                            value,
+                            ", ".join([t.__name__ for t in expected_iter_type]),
+                        )
+                    )
+                else:
+                    msg = (
+                        f'Unable to set "{name}" to "{value}" since each '
[…]
