### 2025-03-12 b322b61
Did simple test defining check_arguments behavior.
