### 2025-03-12 6440f52
Suppressed line expansion when expanding into new line.
