### 2025-03-12 44c941a
Fixed bug with variable definition order.
