### 2025-03-12 3ef8019
Stopped expecting line Expansion with new line in tests.
