### 2025-03-12 fda9553
Added parsed material to tests for syntax tree bugs.
