### 2025-03-12 0e44932
Tried out capturing warnings.
