### 2025-03-12 8c4134f
Combined all LineExpansionWarnings.
