### 2025-03-12 f57ea5f
Made parsed material identitical to simplify tests.
