+++ tests/test_material.py
@@ -48,9 +48,23 @@ class TestMaterial:
     @pytest.fixture
     def parsed_material(_):
         return Material(
-            """m1 1001.00c 0.05
-      1001.04c 0.05
-      1002     0.05"""
+            """M1   1001.00c   0.05
+      1001.04c   0.05
+      1001.80c   0.05
+      1001.04p   0.05
+      1002       0.05
+      1003       0.05
+      90232      0.05
+      90232.701nc 0.05
+      92235      0.05
+      92235.80c  0.05
+      92635.80c  0.05
+      92238      0.05
+      95642      0.05
+      95242      0.05
+      27560.50c  0.05
+      94239      0.05
+      28000.60c  0.05"""
         )
 
     @pytest.fixture
