### 2025-03-12 949f7f4
modularized warning system, and restored previous jank method to pass data.
