### 2025-03-12 cea23b8
Added warning combiner to cell.
