### 2025-03-12 bcac6bf
Restored less janky way of warning data passing.
