### 2025-03-12 c278d44
Updated problem level warning with single expansion warning, and simplified.
