### 2025-03-12 1ea47c6
Removed catch_warnings arguments added in py311
