### 2025-03-12 baa1b99
Added edge case test for shortcuts in materials.
