### 2025-03-12 5cf44c5
Fixed bug with single match in even number sequence
