+++ montepy/input_parser/parser_base.py
@@ -221,12 +221,13 @@ class MCNP_Parser(Parser, metaclass=MetaBuilder):
             sequence.append(p[0])
         else:
             sequence = p[0]
-        if type(p[1]) == syntax_node.ListNode:
-            for node in p[1].nodes:
-                sequence.append(node)
-        else:
-            for idx in range(1, len(p)):
-                sequence.append(p[idx])
+        if len(p) > 1:
+            if type(p[1]) == syntax_node.ListNode:
+                for node in p[1].nodes:
+                    sequence.append(node)
+            else:
+                for idx in range(1, len(p)):
+                    sequence.append(p[idx])
         return sequence
 
     @_("number_phrase", "null_phrase")
