### 2025-03-12 97fcde2
Simplified logic and removed dead code for even num sequence.
