+++ montepy/input_parser/parser_base.py
@@ -216,18 +216,12 @@ class MCNP_Parser(Parser, metaclass=MetaBuilder):
         """
         if not hasattr(p, "even_number_sequence"):
             sequence = syntax_node.ListNode("number sequence")
-            if type(p[0]) == syntax_node.ListNode:
-                return p[0]
             sequence.append(p[0])
         else:
             sequence = p[0]
         if len(p) > 1:
-            if type(p[1]) == syntax_node.ListNode:
-                for node in p[1].nodes:
-                    sequence.append(node)
-            else:
-                for idx in range(1, len(p)):
-                    sequence.append(p[idx])
+            for idx in range(1, len(p)):
+                sequence.append(p[idx])
         return sequence
 
     @_("number_phrase", "null_phrase")
