### 2025-03-14 0189cc3
Updated is_adjective explanation.
