### 2025-03-14 6730b49
Clarified MCNP_classes and fixed syntax issues.
