+++ montepy/mcnp_object.py
@@ -4,8 +4,6 @@ from abc import ABC, ABCMeta, abstractmethod
 import copy
 import functools
 import itertools as it
-import json
-import numpy as np
 import sys
 import textwrap
 from typing import Union
