### 2025-03-14 37bd684
Simplified logic with @tjlaboss's help.
