+++ montepy/mcnp_object.py
@@ -243,28 +243,28 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
         return lines
 
     def _flush_line_expansion_warning(self, lines, ws):
-        message = None
-        if len(ws) > 0:
-            message = f"""The input had a value expand that may change formatting.
-    The new input was:\n\n"""
-            for line in lines:
-                message += f"    {line}"
-            width = 15
-            message += f"\n\n    {'old value': ^{width}s} {'new value': ^{width}s}"
-            message += f"\n    {'':-^{width}s} {'':-^{width}s}\n"
-            olds = []
-            news = []
-            for w in ws:
-                warning = w.message
-                formatter = f"    {{w.og_value: >{width}}} {{w.new_value: >{width}}}\n"
-                message += formatter.format(w=warning)
-                olds.append(warning.og_value)
-                news.append(warning.new_value)
-            if message is not None:
-                warning = LineExpansionWarning(message)
-                warning.olds = olds
-                warning.news = news
-                warnings.warn(warning, stacklevel=4)
+        if not ws:
+            return
+        message = f"""The input had a value expand that may change formatting.
+The new input was:\n\n"""
+        for line in lines:
+            message += f"    {line}"
+        width = 15
+        message += f"\n\n    {'old value': ^{width}s} {'new value': ^{width}s}"
+        message += f"\n    {'':-^{width}s} {'':-^{width}s}\n"
+        olds = []
+        news = []
+        for w in ws:
+            warning = w.message
+            formatter = f"    {{w.og_value: >{width}}} {{w.new_value: >{width}}}\n"
+            message += formatter.format(w=warning)
+            olds.append(warning.og_value)
+            news.append(warning.new_value)
+        if message is not None:
+            warning = LineExpansionWarning(message)
+            warning.olds = olds
+            warning.news = news
+            warnings.warn(warning, stacklevel=4)
 
     @property
     def comments(self) -> list[PaddingNode]:
