+++ doc/source/changelog.rst
@@ -27,6 +27,7 @@ MontePy Changelog
 * Particle type exceptions are now warnings, not errors (:issue:`381`).
 * Added :func:`~montepy.data_inputs.material.Material.clear` to ``Material`` to clear out all nuclides (:issue:`665`).
 * Allow any ``Real`` type for floating point numbers and any ``Integral`` type for integer numbers during type enforcement (:issue:`679`).
+* Avoided multiple ``LineExpansionWarnings`` coming from the same object on export (:issue:`198`).
 
 **Bugs Fixed**
 
