### 2025-03-16 573f0c3
Made export verifier, and replicated bug #697
