### 2025-03-16 e92c136
Added #697 to the changelog.
