### 2025-03-16 8b5344d
Promoted Lattice enum to top level.
