### 2025-03-16 cfa9097
Made cell verifier.
