+++ tests/test_cell_problem.py
@@ -336,3 +336,57 @@ def test_bad_setattr():
         cell.nuber = 5
     cell._nuber = 5
     assert cell._nuber == 5
+
+
+def verify_export(cell):
+    output = cell.format_for_mcnp_input((6, 3, 0))
+    print("cell output", output)
+    new_cell = montepy.Cell("\n".join(output))
+    for attr in {
+        "old_number",
+        "old_mat_number",
+        "old_universe_number",
+        "lattice",
+        "mass_density",
+        "atom_density",
+        "is_atom_dens",
+    }:
+        try:
+            old_attr = getattr(cell, attr)
+            new_attr = getattr(new_cell, attr)
+            # jank override
+            if attr == "old_universe_number" and cell.universe:
+                old_attr = cell.universe.number
+        except AttributeError as e:
+            if "density" not in attr:
+                raise e
+            else:
+                continue
+        print(f"attr: {attr}, old: {old_attr}, new: {new_attr}")
+        if old_attr is not None:
+            if isinstance(old_attr, float):
+                assert old_attr == pytest.approx(new_attr)
+            else:
+                assert old_attr == new_attr
+        else:
+            assert new_attr is None
+    for attr in {
+        "hidden_transform",
+        "multiple_universes",
+        "old_universe_number",
+        "old_universe_numbers",
+        "transform",
+    }:
+        old_attr = getattr(cell.fill, attr)
+        new_attr = getattr(new_cell.fill, attr)
+        # jank override
+        if attr == "old_universe_number" and cell.fill.universe:
+            old_attr = cell.fill.universe.number
+        print(f"fill attr: {attr}, old: {old_attr}, new: {new_attr}")
+        if old_attr is not None:
+            if isinstance(old_attr, float):
+                assert old_attr == pytest.approx(new_attr)
+            else:
+                assert old_attr == new_attr
+        else:
+            assert new_attr is None
