### 2025-03-17 ab27c14
Clarified PyNE Can't read MCNP input files.
