### 2025-03-17 dcca284
Updated tests for changing print_data_block default.
