### 2025-03-17 a50b391
Updated docs with new defaults.
