### 2025-03-17 8d28ba2
Added #699 to the changelog.
