### 2025-03-17 0fa698d
Simplified dictionary logic.
