+++ montepy/input_parser/surface_parser.py
@@ -1,4 +1,4 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from montepy.input_parser.parser_base import MCNP_Parser
 from montepy.input_parser.tokens import SurfaceLexer
 from montepy.input_parser import syntax_node
