+++ montepy/__init__.py
@@ -1,5 +1,4 @@
 # Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 """MontePy is a library for reading, editing, and writing MCNP input files.
 
 This creates a semantic understanding of the MCNP input file.
