### 2025-03-18 3400f1c
Simplified logic and made clearer.
