+++ tests/test_surfaces.py
@@ -1,5 +1,6 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from unittest import TestCase
+import pytest
 
 import montepy
 
@@ -312,3 +313,15 @@ class testSurfaces(TestCase):
         # test length issues
         with self.assertRaises(ValueError):
             surf.coordinates = [3, 4, 5]
+
+
+@pytest.mark.parametrize(
+    "surf_str", ["1 PZ 0.0", "1 SO 1.0", "1 CZ 9.0", "4 C/z 5.0 0 3"]
+)
+def test_surface_clone(surf_str):
+    prob = montepy.MCNP_Problem("")
+    surf = surface_builder(surf_str)
+    prob.surfaces.append(surf)
+    new_surf = surf.clone()
+    assert surf.surface_type == new_surf.surface_type
+    assert surf.surface_constants == new_surf.surface_constants
