### 2025-03-18 f70a5f9
Look for matching parent class in cloning collection.
