### 2025-03-18 d77a182
Added #704 to the changelog.
