### 2025-03-18 ea4b0de
Set default value to 0 for universe.
