### 2025-03-19 603f2bf
Simplified logic.
