### 2025-03-19 e0bea2b
Updating changelog to next version: 0.5.5
