### 2025-03-20 026fb10
Caught case where values are unset.
