### 2025-03-20 #6977 @MicahGale
MontePy is now on [Zenodo](https://zenodo.org/records/15058701): [![DOI](https://zenodo.org/badge/DOI/10.11578/dc.20240115.1.svg)](https://doi.org/10.11578/dc.20240115.1)
