### 2025-03-20 b52fb82
Added PyOpenSci and Zenodo badges.
