### 2025-03-21 6232eed
Tested is_reflecting with lots of leading spaces.
