+++ tests/test_surfaces.py
@@ -1,5 +1,7 @@
 # Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
+import io
 from unittest import TestCase
+from pathlib import Path
 import pytest
 
 import montepy
@@ -128,25 +130,6 @@ class testSurfaces(TestCase):
         surf._surface_type = SurfaceType.P
         with self.assertRaises(montepy.errors.IllegalState):
             surf.validate()
-
-    def test_surface_is_reflecting_setter(self):
-        for in_str, expected in [("1 PZ 0.0", True), ("*1 PZ 0.0", False)]:
-            surf = Surface(in_str)
-            surf.is_reflecting = expected
-            assert surf.is_reflecting == expected
-            self.verify_export(surf)
-            with self.assertRaises(TypeError):
-                surf.is_reflecting = 1
-
-    # @pytest.mark.parametrize("in_str, expected", [("1 PZ 0.0", True), ("+1 PZ 0.0", False)])
-    def test_surface_is_white_bound_setter(self):
-        for in_str, expected in [("1 PZ 0.0", True), ("+1 PZ 0.0", False)]:
-            surf = Surface(in_str)
-            surf.is_white_boundary = expected
-            assert surf.is_white_boundary == expected
-            self.verify_export(surf)
-            with self.assertRaises(TypeError):
-                surf.is_white_boundary = 1
 
     def test_surface_constants_setter(self):
         in_str = "1 PZ 0.0"
@@ -314,24 +297,36 @@ class testSurfaces(TestCase):
         with self.assertRaises(ValueError):
             surf.coordinates = [3, 4, 5]
 
-    def verify_export(_, surf):
-        output = surf.format_for_mcnp_input((6, 3, 0))
-        print("Surface output", output)
-        new_surf = Surface("\n".join(output))
-        assert surf.number == new_surf.number, "Material number not preserved."
-        assert len(surf.surface_constants) == len(
-            new_surf.surface_constants
-        ), "number of surface constants not kept."
-        for old_const, new_const in zip(
-            surf.surface_constants, new_surf.surface_constants
-        ):
-            assert old_const == pytest.approx(new_const)
-        assert surf.is_reflecting == new_surf.is_reflecting
-        assert surf.is_white_boundary == new_surf.is_white_boundary
-        if surf.old_periodic_surface:
-            assert surf.old_periodic_surface == new_surf.old_periodic_surface
-        if surf.old_transform_number:
-            assert surf.old_transform_number == new_surf._old_transform_number
+
+def verify_export(surf):
+    output = surf.format_for_mcnp_input((6, 3, 0))
+    print("Surface output", output)
+    new_surf = Surface("\n".join(output))
+    verify_equiv_surf(surf, new_surf)
+
+
+def verify_equiv_surf(surf, new_surf):
+    assert surf.number == new_surf.number, "Material number not preserved."
+    assert len(surf.surface_constants) == len(
+        new_surf.surface_constants
+    ), "number of surface constants not kept."
+    for old_const, new_const in zip(surf.surface_constants, new_surf.surface_constants):
+        assert old_const == pytest.approx(new_const)
+    assert surf.is_reflecting == new_surf.is_reflecting
+    assert surf.is_white_boundary == new_surf.is_white_boundary
+    if surf.old_periodic_surface:
+        assert surf.old_periodic_surface == new_surf.old_periodic_surface
+    if surf.old_transform_number:
[…]
