### 2025-03-21 aac47f4
Made mechanism to generate default never pad node.
