### 2025-03-21 1a0f799
Found sneaky way never_pad argument is missed.
