### 2025-03-22 c87e7f5
Played around with function inspection.
