### 2025-03-23 5d87253
chars are strings... whoops.
