### 2025-03-23 d34102f
Made function check wrapper.
