+++ pyproject.toml
@@ -7,12 +7,12 @@ description = "A library for reading, editing, and writing MCNP input files"
 readme = "README.md"
 requires-python = ">=3.9"
 maintainers = [
-	{name = "Micah Gale", email = "mgale@montepy.org"}
+    {name = "Micah Gale", email = "mgale@montepy.org"}
 ]
 authors = [
-	{name = "Micah Gale", email = "mgale@montepy.org"},
-	{name = "Travis Labossiere-Hickman", email = "Travis.LabossiereHickman@inl.gov"},
-	{name = "Brenna Carbno", email="brenna.carbno@inl.gov"}
+    {name = "Micah Gale", email = "mgale@montepy.org"},
+    {name = "Travis Labossiere-Hickman", email = "Travis.LabossiereHickman@inl.gov"},
+    {name = "Brenna Carbno", email="brenna.carbno@inl.gov"}
 ]
 keywords = ["MCNP", "neutronics", "imcnp", "input file", "monte carlo", "radiation transport"]
 license = {file="LICENSE"}
@@ -34,8 +34,8 @@ classifiers = [
        "Programming Language :: Python :: 3.13"
 ]
 dependencies = [
-	"numpy>=1.18",
-	"sly>= 0.4, <=0.5"
+    "numpy>=1.18",
+    "sly>= 0.4, <=0.5"
 ]
 
 [project.optional-dependencies]
@@ -49,22 +49,22 @@ test = [
 ]
 # This is needed for a sphinx bug. See #414.
 doc = [
-	"sphinx>=7.4.0", 
-	"sphinxcontrib-apidoc", 
-	"pydata_sphinx_theme",
-	"sphinx-favicon",
-	"sphinx-copybutton",
-    	"sphinx_autodoc_typehints",
-	"autodocsumm",
+    "sphinx>=7.4.0", 
+    "sphinxcontrib-apidoc", 
+    "pydata_sphinx_theme",
+    "sphinx-favicon",
+    "sphinx-copybutton",
+    "sphinx_autodoc_typehints",
+    "autodocsumm",
 ]
 format = ["black~=25.1"]
 build = [
-	"build",
-	"setuptools>=64.0.0",
-	"setuptools-scm>=8",
+    "build",
+    "setuptools>=64.0.0",
+    "setuptools-scm>=8",
 ]
 develop = [
-	"montepy[test,doc,format]",
+    "montepy[test,doc,format]",
 ]
 demos = ["jupyter"]
 demo-test = ["montepy[demos]", "papermill"]
@@ -98,8 +98,8 @@ precision = 2
 show_missing = true
 fail_under = 90.0
 exclude_also = [
-	"\\s+@abstractmethod\\s*",
-	"pass",
+    "\\s+@abstractmethod\\s*",
+    "pass",
 ]
 
 [tool.pytest.ini_options]
