### 2025-03-23 c4e396c
Tried to be exhaustive test cases into check value system.
