+++ tests/test_check_value.py
@@ -6,8 +6,45 @@ import montepy._check_value as cv
 
 
 @cv.check_arguments
-def dummy_func(a: str):
+def simple_args(a: str):
     pass
+
+
+def var_args(a: str):
+    @cv.check_arguments
+    def ret(a: str, *args: str):
+        pass
+
+    return ret("1", a, a)
+
+
+def kwargs(a: str):
+    @cv.check_arguments
+    def ret(**kwargs: str):
+        pass
+
+    return ret(foo=a)
+
+
+@cv.check_arguments
+def defaults(a: str, *args: str, b: str = None):
+    pass
+
+
+def kw_only(a: str):
+    return defaults("1", "1", "1", b=a)
+
+
+def defaults_used(a: str):
+    return defaults("1")
+
+
+def kw_defaults(a: str):
+    @cv.check_arguments
+    def ret(a: str, *args: str, b: str = None, **kwargs: str):
+        pass
+
+    return ret(foo=a)
 
 
 binary = st.binary()
@@ -19,12 +56,22 @@ it = st.integers()
 no = st.none()
 
 
-@given(st.one_of(binary, boolean, dt, fl, it, no))
-def test_dummy_bad_type(val):
+@given(
+    st.one_of(binary, boolean, dt, fl, it, no),
+    st.sampled_from(
+        [simple_args, var_args, kwargs, kw_only, kw_defaults, defaults_used]
+    ),
+)
+def test_dummy_bad_type(val, func):
     with pytest.raises(TypeError):
-        dummy_func(val)
+        func(val)
 
 
-@given(st.text())
-def test_dummy_good_type(val):
-    dummy_func(val)
+@given(
+    st.one_of(binary, boolean, dt, fl, it, no),
+    st.sampled_from(
+        [simple_args, var_args, kwargs, kw_only, kw_defaults, defaults_used]
+    ),
+)
+def test_dummy_good_type(val, func):
+    func(val)
