### 2025-03-23 8ce1ed3
Tested checking with annotations enabled.
