+++ tests/test_importance.py
@@ -1,180 +1,320 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 import montepy
+import os, pytest
 from montepy.cell import Cell
 from montepy.particle import Particle
 from montepy.data_inputs.importance import Importance
 from montepy.errors import *
 from montepy.input_parser import mcnp_input, block_type
-import os
-import io
-import pytest
 
 
-def test_importance_init_cell():
-    # test_normal cell init
-    in_str = "1 0 -1 IMP:N,P=1"
-    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
-    cell = Cell(card)
-    assert cell.importance.neutron == 1.0
-    assert cell.importance.photon == 1.0
-    assert cell.importance.alpha_particle == 0.0
-    assert cell.importance.all is None
-    assert cell.importance.in_cell_block
-    # test non-number imp
-    in_str = "1 0 -1 IMP:N,P=h"
-    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
-    with pytest.raises(ValueError):
-        cell = Cell(card)
-    # test negative imp
-    in_str = "1 0 -1 IMP:N,P=-2"
-    card = mcnp_input.Input([in_str], block_type.BlockType.CELL)
-    with pytest.raises(ValueError):
-        cell = Cell(card)
+def create_cell_from_input(in_str, block=block_type.BlockType.CELL):
+    """Helper to create a Cell object from a given input string."""
+    card = mcnp_input.Input([in_str], block)
+    return Cell(card)
 
 
-def test_importance_init_data():
-    in_str = "IMP:N,P 1 0"
+# Parameterized test for valid importance parsing
+# Test cases:
+# 1. Combined importance for neutrons and photons (IMP:N,P=1)
+#    - Tests multiple particle types in single declaration
+#    - Verifies default values for unspecified particles
+# 2. Separate importance for electrons and protons (IMP:E=1 IMP:H=1)
+#    - Tests individual particle importance declarations
+@pytest.mark.parametrize(
+    "in_str,expected",
+    [
+        (
+            "1 0 -1 IMP:N,P=1",
+            {
+                "neutron": 1.0,
+                "photon": 1.0,
+                "all": None,
+                "alpha_particle": 0.0,
+                "in_cell_block": True,
+            },
+        ),
+        (
+            "1 0 -1 IMP:E=1 IMP:H=1",
+            {"electron": 1.0, "proton": 1.0},
+        ),
+    ],
+)
+def test_importance_parsing_from_cell(in_str, expected):
+    """
+    Testing all examples of importance input
+    """
+    cell = create_cell_from_input(in_str)
+    for attr, value in expected.items():
+        actual = getattr(cell.importance, attr)
+        assert actual == value, f"Expected {attr}={value}, got {actual}"
+
+
+# Parameterized test for invalid importance values
+# Test cases:
+# 1. Non-numeric value 'h' for importance
+#    - Verifies rejection of non-numeric importance values
+#    - Should raise ValueError
+# 2. Negative value -2 for importance
+#    - Verifies rejection of negative importance values
+#    - Should raise ValueError since importance must be non-negative
+@pytest.mark.parametrize(
+    "in_str,expected_exception",
+    [
+        ("1 0 -1 IMP:N,P=h", ValueError),  # non-numeric value should raise ValueError
+        ("1 0 -1 IMP:N,P=-2", ValueError),  # negative value should raise ValueError
+    ],
+)
+def test_invalid_importance_parsing_from_cell(in_str, expected_exception):
+    """
+    Testing invalid importance values
+    Every test shuould Raises ValueError
+    """
+    with pytest.raises(expected_exception):
[…]
