+++ pyproject.toml
@@ -13,6 +13,7 @@ authors = [
 	{name = "Micah Gale", email = "mgale@montepy.org"},
 	{name = "Travis Labossiere-Hickman", email = "Travis.LabossiereHickman@inl.gov"},
 	{name = "Brenna Carbno", email="brenna.carbno@inl.gov"}
+	{name = "Benjaminas Marcinkevicius", email="BenjaminasDev@outlook.com"}
 ]
 keywords = ["MCNP", "neutronics", "imcnp", "input file", "monte carlo", "radiation transport"]
 license = {file="LICENSE"}
@@ -64,7 +65,7 @@ build = [
 	"setuptools-scm>=8",
 ]
 develop = [
-	"montepy[test,doc,format,demo-test]",
+	"montepy[test,doc,format]",
 ]
 demos = ["jupyter"]
 demo-test = ["montepy[demos]", "papermill"]
