+++ tests/test_importance.py
@@ -14,16 +14,11 @@ def create_cell_from_input(in_str, block=block_type.BlockType.CELL):
     return Cell(card)
 
 
-# Parameterized test for valid importance parsing
-# Test cases:
-# 1. Combined importance for neutrons and photons (IMP:N,P=1)
-#    - Tests multiple particle types in single declaration
-#    - Verifies default values for unspecified particles
-# 2. Separate importance for electrons and protons (IMP:E=1 IMP:H=1)
-#    - Tests individual particle importance declarations
+# Parameterized test for valid and invalid importance parsing
 @pytest.mark.parametrize(
-    "in_str,expected",
+    "in_str, expected, error",
     [
+        # Valid cases
         (
             "1 0 -1 IMP:N,P=1",
             {
@@ -33,65 +28,70 @@ def create_cell_from_input(in_str, block=block_type.BlockType.CELL):
                 "alpha_particle": 0.0,
                 "in_cell_block": True,
             },
+            None,
         ),
         (
             "1 0 -1 IMP:E=1 IMP:H=1",
             {"electron": 1.0, "proton": 1.0},
+            None,
         ),
+        # Error cases
+        ("1 0 -1 IMP:N,P=h", None, ValueError),  # non-numeric value
+        ("1 0 -1 IMP:N,P=-2", None, ValueError),  # negative value
+        ("1 0 -1 IMP:N,xx=2", None, ParsingError),  # invalid particle type
     ],
 )
-def test_importance_parsing_from_cell(in_str, expected):
+def test_importance_parsing_from_cell(in_str, expected, error):
+    """Test importance parsing from cell input string.
+    
+    Tests both valid and invalid cases:
+    - Valid: Verifies parsed values match expected values
+    - Invalid: Verifies appropriate errors are raised
     """
-    Testing all examples of importance input
-    """
-    cell = create_cell_from_input(in_str)
-    for attr, value in expected.items():
-        actual = getattr(cell.importance, attr)
-        assert actual == value, f"Expected {attr}={value}, got {actual}"
-
-
-# Parameterized test for invalid importance values
-# Test cases:
-# 1. Non-numeric value 'h' for importance
-#    - Verifies rejection of non-numeric importance values
-#    - Should raise ValueError
-# 2. Negative value -2 for importance
-#    - Verifies rejection of negative importance values
-#    - Should raise ValueError since importance must be non-negative
-@pytest.mark.parametrize(
-    "in_str,expected_exception",
-    [
-        ("1 0 -1 IMP:N,P=h", ValueError),  # non-numeric value should raise ValueError
-        ("1 0 -1 IMP:N,P=-2", ValueError),  # negative value should raise ValueError
-    ],
-)
-def test_invalid_importance_parsing_from_cell(in_str, expected_exception):
-    """
-    Testing invalid importance values
-    Every test shuould Raises ValueError
-    """
-    with pytest.raises(expected_exception):
-        create_cell_from_input(in_str)
+    if error is not None:
+        with pytest.raises(error):
+            create_cell_from_input(in_str)
+    else:
+        cell = create_cell_from_input(in_str)
+        for attr, value in expected.items():
+            actual = getattr(cell.importance, attr)
+            assert actual == value, f"Expected {attr}={value}, got {actual}"
 
 
 # Valid data case: the card should be parsed correctly
 @pytest.mark.parametrize(
-    "in_str,expected_neutron,expected_photon",
+    "in_str, expected_values",
     [
-        ("IMP:N,P 1 0", [1.0, 0.0], [1.0, 0.0]),
+        (
+            "IMP:N,P 1 0",
+            {
[…]
