### 2025-03-24 79e21bd
added test on default importance value
