+++ .github/workflows/main.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13", "3.14-dev"]
 
     steps:
       - uses: actions/checkout@v4
@@ -54,8 +54,8 @@ jobs:
       actions: write
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
-        numpy-version: ["1.0", "2.0"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13", "3.14-dev"]
+        numpy-version: ["2.0"]
         sly-version: ["0.5"]
         include:
           - python-version: "3.11"
