+++ .github/workflows/main.yml
@@ -61,6 +61,12 @@ jobs:
           - python-version: "3.11"
             numpy-version: "2.0"
             sly-version: "0.4"
+          - python-version: "3.13"
+            numpy-version: "1.0"
+            sly-version: "0.4"
+          - python-version: "3.9"
+            numpy-version: "1.0"
+            sly-version: "0.5"
     
     steps:
       - uses: actions/checkout@v4
