### 2025-03-24 21c39e3
using correct float comparison with pytest.approx
