+++ tests/test_universe.py
@@ -1,4 +1,6 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+from hypothesis import given, strategies as st
+import pytest
 from unittest import TestCase
 
 import copy
@@ -388,3 +390,61 @@ class TestFill(TestCase):
         fill2 = Fill(card)
         with self.assertRaises(MalformedInputError):
             fill1.merge(fill2)
+
+    @given(
+        indices=st.lists(st.integers(), min_size=3, max_size=3),
+        width=st.lists(st.integers(1), min_size=3, max_size=3),
+    )
+    def test_fill_index_setter(self, indices, width):
+        fill = self.simple_fill.clone()
+        fill.multiple_universes = True
+        fill.min_index = indices
+        end = np.array(indices) + np.array(width)
+        fill.max_index = end
+        assert fill.min_index == indices
+        assert fill.max_index == end
+
+    def test_fill_index_bad_setter(self):
+        fill = self.simple_fill
+        with pytest.raises(TypeError):
+            fill.min_index = "hi"
+        with pytest.raises(TypeError):
+            fill.max_index = "hi"
+        with pytest.raises(TypeError):
+            fill.min_index = ["hi"]
+        with pytest.raises(TypeError):
+            fill.max_index = ["hi"]
+        with pytest.raises(ValueError):
+            fill.min_index = [1]
+        with pytest.raises(ValueError):
+            fill.max_index = [1]
+
+    @given(
+        universes=st.lists(st.integers(0), min_size=1, max_size=10),
+        y_len=st.integers(1, 10),
+        z_len=st.integers(1, 10),
+    )
+    def test_fill_multi_unis(self, universes, y_len, z_len):
+        fill = self.simple_fill.clone()
+        universes = np.array([[Universe(u) for u in universes] * y_len] * z_len)
+        fill.multiple_universes = True
+        fill.universes = universes
+        assert fill.universes == universes
+        assert fill.min_index == np.array([0, 0, 0])
+        assert fill.max_index == np.array(universes.shape)
+        self.verify_export(fill)
+
+    def verify_export(self, fill):
+        output = fill.format_for_mcnp_input((6,3,0))
+        print(output)
+        cell = montepy.Cell("1 0 -2 "+ "\n".join(output))
+        for attr in ["multiple_universes", "old_universe_numbers", "old_universe_number"]:
+            old_val = getattr(fill, attr)
+            if "old" in attr:
+                if "s":
+                    old_val = getattr(fill, "universes")
+                else:
+                    old_val = getattr(fill, "universe")
+            new_val = getattr(fill, attr)
+            print(attr, old_val, new_val)
+            assert old_val == new_val
