### 2025-03-25 d756599
Made min_index and max_index settable.
