### 2025-03-25 ea67ac2
Covered most edge cases with argument for check args.
