### 2025-03-25 c585aff
Made decorator wrapper.
