### 2025-03-26 9e7bb75
Covered more error paths.
