+++ montepy/mcnp_object.py
@@ -118,11 +118,6 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
                     input.split("\n"), self._BLOCK_TYPE
                 )
             try:
-                try:
-                    parser.restart()
-                # raised if restarted without ever parsing
-                except AttributeError as e:
-                    pass
                 self._tree = parser.parse(input.tokenize(), input)
                 self._input = input
             except ValueError as e:
