### 2025-03-26 #7951 @editorialbot
[REVIEW]: MontePy: a Python library for reading, editing, and writing MCNP input files.
**Submitting author:** @micahgale (<a href="http://orcid.org/0000-0001-6451-4818">Micah D. Gale</a>)
**Repository:** https://github.com/idaholab/MontePy
**Branch with paper.md** (empty if default branch): joss_submit
**Version:** v0.5.5
