### 2025-03-26 #7951 @kellyrowland
hi @kyleniemeyer we have completed the pyOpenSci review of this package at https://github.com/pyOpenSci/software-submission/issues/205
