### 2025-03-27 3f1022a
corrected error type raised
