+++ tests/test_importance.py
@@ -254,7 +254,7 @@ class TestImportance:
         problem = montepy.mcnp_problem.MCNP_Problem("foo")
         cell.link_to_problem(problem)
         # test problem mode enforcement
-        with pytest.raises(ValueError):
+        with pytest.raises(ParticleTypeNotInProblem):
             cell.importance.photon = 1.0
         # test wrong type
         with pytest.raises(TypeError):
