### 2025-03-27 0caa7ab
moved the format_default into __str__ for brevity
