### 2025-03-27 1c198bf
refactor to utilise Importance(in_str) initialisation
