### 2025-03-27 886e7d6
moved all functions into the class and reformatted
