+++ tests/test_importance.py
@@ -10,134 +10,116 @@ from montepy.errors import *
 from montepy.input_parser import mcnp_input, block_type
 
 
-def create_cell_from_input(in_str, block=block_type.BlockType.CELL):
-    """Helper to create a Cell object from a given input string."""
-    card = mcnp_input.Input([in_str], block)
-    return Cell(card)
-
-
-# Parameterized test for valid and invalid importance parsing
-@pytest.mark.parametrize(
-    "in_str, expected, error",
-    [
-        # Valid cases
-        (
-            "1 0 -1 IMP:N,P=1",
-            {
-                "neutron": 1.0,
-                "photon": 1.0,
-                "all": None,
-                "alpha_particle": 0.0,
-                "in_cell_block": True,
-            },
-            None,
-        ),
-        (
-            "1 0 -1 IMP:E=1 IMP:H=1",
-            {"electron": 1.0, "proton": 1.0},
-            None,
-        ),
-        (
-            "1 0 -1",
-            {"neutron": 0.0},
-            None,
-        ),  # default neutron importance when nothing is set
-        # Error cases
-        ("1 0 -1 IMP:N,P=h", None, ValueError),  # non-numeric value
-        ("1 0 -1 IMP:N,P=-2", None, ValueError),  # negative value
-        ("1 0 -1 IMP:N,xx=2", None, ParsingError),  # invalid particle type
-    ],
-)
-def test_importance_parsing_from_cell(in_str, expected, error):
-    """Test importance parsing from cell input string.
-
-    Tests both valid and invalid cases:
-    - Valid: Verifies parsed values match expected values
-    - Invalid: Verifies appropriate errors are raised
-    """
-    if error is not None:
-        with pytest.raises(error):
-            create_cell_from_input(in_str)
-    else:
-        cell = create_cell_from_input(in_str)
-        for attr, value in expected.items():
-            actual = getattr(cell.importance, attr)
-            assert actual == pytest.approx(
-                value
-            ), f"Expected {attr}={value}, got {actual}"
-
-
-# Valid data case: the card should be parsed correctly
-@pytest.mark.parametrize(
-    "in_str, expected_values",
-    [
-        (
-            "IMP:N,P 1 0",
-            {
-                Particle.NEUTRON: [1.0, 0.0],
-                Particle.PHOTON: [1.0, 0.0],
-            },
-        ),
-        (
-            "IMP:N,P,E 1 0 2",
-            {
-                Particle.NEUTRON: [1.0, 0.0, 2.0],
-                Particle.PHOTON: [1.0, 0.0, 2.0],
-                Particle.ELECTRON: [1.0, 0.0, 2.0],
-            },
-        ),
-    ],
-)
-def test_importance_init_data_valid(in_str, expected_values):
-    """Test importance data initialization for multiple particles.
-
-    Args:
-        in_str: Input string containing importance definitions
-        expected_values: Dictionary mapping particles to their expected importance values
-    """
-    imp = Importance(in_str)  # Updated: directly pass the string instead of creating card
-
-    for particle, value in expected_values.items():
-        actual = [val.value for val in imp._particle_importances[particle]["data"]]
-        assert actual == pytest.approx(
-            value
-        ), f"For {particle.name}, expected {value}, got {actual}"
-
-
-# Error cases: each input should raise an exception, optionally with additional keyword arguments.
-@pytest.mark.parametrize(
-    "in_str, kwargs, expected_exception",
-    [
-        ("IMP:N,P 1 h", {}, ValueError),  # non-numeric importance
-        ("IMP:N,P 1 -2", {}, ValueError),  # negative importance
[…]
