### 2025-03-27 a1fd4c6
Update doc/source/changelog.rst
Co-authored-by: Micah Gale <mgale@fastmail.com>
