+++ montepy/mcnp_problem.py
@@ -8,6 +8,7 @@ import threading
 import os
 import sys
 import queue
+import time
 from typing import Any
 import warnings
 
@@ -386,12 +387,29 @@ class MCNP_Problem:
             if input is None:
                 to_parse.task_done()
                 to_parse.cancel_join_thread()
+                parsed.put(cls._PriotizedItem(idx, None))
+                parsed.close()
                 parsed.cancel_join_thread()
                 return
             obj_parser = cls._OBJ_MATCHER[input.block_type]
             obj = obj_parser(input)
             parsed.put_nowait(cls._PriotizedItem(idx, obj))
             to_parse.task_done()
+
+    @staticmethod
+    def _consume_parsed_queue(num_threads, parsed, output):
+        threads_died = 0
+        while True:
+            obj = parsed.get()
+            parsed.task_done()
+            print(f"obj: {obj}, {parsed.qsize()}, {parsed.empty()}, {parsed.full()}")
+            if obj.item is None:
+                threads_died += 1
+                if threads_died == num_threads:
+                    parsed.cancel_join_thread()
+                    return
+            else:
+                output[obj.priority] = obj.item
 
     def parse_input(self, check_input=False, replace=True, num_threads=None):
         """Semantically parses the MCNP file provided to the constructor.
@@ -404,57 +422,77 @@ class MCNP_Problem:
         replace : bool
             replace all non-ASCII characters with a space (0x20)
         """
-        trailing_comment = None
-        last_obj = None
-        last_block = None
-        OBJ_MATCHER = {
-            block_type.BlockType.CELL: (Cell, self._cells),
-            block_type.BlockType.SURFACE: (
-                surface_builder.parse_surface,
-                self._surfaces,
-            ),
-            block_type.BlockType.DATA: (parse_data, self._data_inputs),
-        }
-        to_parse = mp.JoinableQueue()
-        parsed = mp.Queue()
-        errors = mp.Queue()
-        if num_threads is None:
-            num_threads = os.cpu_count()
-        processes = []
-        for _ in range(num_threads):
-            proc = mp.Process(
-                target=self._parse_as_thread, args=[to_parse, parsed, errors]
+        try:
+            start = time.time()
+            trailing_comment = None
+            last_obj = None
+            last_block = None
+            OBJ_MATCHER = {
+                block_type.BlockType.CELL: (Cell, self._cells),
+                block_type.BlockType.SURFACE: (
+                    surface_builder.parse_surface,
+                    self._surfaces,
+                ),
+                block_type.BlockType.DATA: (parse_data, self._data_inputs),
+            }
+            to_parse = mp.JoinableQueue()
+            parsed_q = mp.JoinableQueue()
+            errors = mp.Queue()
+            parsed = {}
+            if num_threads is None:
+                num_threads = os.cpu_count() - 1
+            processes = []
+            for _ in range(num_threads):
+                proc = mp.Process(
+                    target=self._parse_as_thread, args=[to_parse, parsed_q, errors]
+                )
+                proc.start()
+                processes.append(proc)
+            consumer = threading.Thread(
+                target=self._consume_parsed_queue, args=[num_threads, parsed_q, parsed]
             )
-            proc.start()
-            processes.append(proc)
-        consumer = threading.Thread(
-            target=self._load_parsed_inputs, args=[parsed, check_input]
-        )
+            consumer.start()
+            print(f"Reading started: {time.time() - start}")
+            for i, input in enumerate(
+                input_syntax_reader.read_input_syntax(
+                    self._input_file, self.mcnp_version, replace=replace
+                )
+            ):
+                self._original_inputs.append(input)
+                if i == 0 and isinstance(input, mcnp_input.Message):
+                    self._message = input
 
-        for i, input in enumerate(
-            input_syntax_reader.read_input_syntax(
[…]
