+++ montepy/mcnp_problem.py
@@ -388,28 +388,12 @@ class MCNP_Problem:
                 to_parse.task_done()
                 to_parse.cancel_join_thread()
                 parsed.put(cls._PriotizedItem(idx, None))
-                parsed.close()
-                parsed.cancel_join_thread()
+                # parsed.cancel_join_thread()
                 return
             obj_parser = cls._OBJ_MATCHER[input.block_type]
             obj = obj_parser(input)
-            parsed.put_nowait(cls._PriotizedItem(idx, obj))
+            parsed.put(cls._PriotizedItem(idx, obj))
             to_parse.task_done()
-
-    @staticmethod
-    def _consume_parsed_queue(num_threads, parsed, output):
-        threads_died = 0
-        while True:
-            obj = parsed.get()
-            parsed.task_done()
-            print(f"obj: {obj}, {parsed.qsize()}, {parsed.empty()}, {parsed.full()}")
-            if obj.item is None:
-                threads_died += 1
-                if threads_died == num_threads:
-                    parsed.cancel_join_thread()
-                    return
-            else:
-                output[obj.priority] = obj.item
 
     def parse_input(self, check_input=False, replace=True, num_threads=None):
         """Semantically parses the MCNP file provided to the constructor.
@@ -440,7 +424,7 @@ class MCNP_Problem:
             errors = mp.Queue()
             parsed = {}
             if num_threads is None:
-                num_threads = os.cpu_count() - 1
+                num_threads = os.cpu_count()
             processes = []
             for _ in range(num_threads):
                 proc = mp.Process(
@@ -448,46 +432,39 @@ class MCNP_Problem:
                 )
                 proc.start()
                 processes.append(proc)
-            consumer = threading.Thread(
-                target=self._consume_parsed_queue, args=[num_threads, parsed_q, parsed]
-            )
-            consumer.start()
             print(f"Reading started: {time.time() - start}")
-            for i, input in enumerate(
+            input_iter = iter(
                 input_syntax_reader.read_input_syntax(
                     self._input_file, self.mcnp_version, replace=replace
                 )
-            ):
-                self._original_inputs.append(input)
-                if i == 0 and isinstance(input, mcnp_input.Message):
-                    self._message = input
-
-                elif isinstance(input, mcnp_input.Title) and self._title is None:
-                    self._title = input
-
-                elif isinstance(input, mcnp_input.Input):
-                    to_parse.put((i, input))
+            )
+            input = next(input_iter)
+            self._original_inputs.append(input)
+            if isinstance(input, mcnp_input.Message):
+                self._message = input
+                input = next(input_iter)
+            self._title = input
+            for i, input in enumerate(input_iter):
+                to_parse.put((i, input))
             print(f"Finished file reading: {time.time() - start}")
             for _ in processes:
                 to_parse.put((-1, None))
             to_parse.close()
-            to_parse.join_thread()
             print(f"to_parse queue closed: {time.time() - start}")
             print(
                 "\n\n\n\n\n************************** CLOSED QUEUE ****************************\n\n\n\n\n\n"
             )
+            print(f"parsers done: {time.time() - start}")
+            self._load_parsed_inputs(num_threads, parsed_q, check_input)
+            to_parse.join_thread()
+            parsed_q.join()
             for process in processes:
                 process.join()
-            consumer.join()
-            print(f"parsers done: {time.time() - start}")
-            print(f"Consumer done: {time.time() - start}")
-            #   self._load_parsed_inputs(parsed, check_input)
             print(f" Post processed: {time.time() - start}.")
             while not errors.empty():
                 e = errors.get_nowait()
                 raise e
         finally:
-            consumer.join(1)
             try:
[…]
