### 2025-03-28 7b5b78b
Made the numpy use in fill easier to read.
