### 2025-03-29 3032177
Tested new line line expansion bug, #721.
