### 2025-03-29 8767afe
Summarized new line additions warnings to avoid formatting list.
