### 2025-03-31 d18c81a
Created test to replicate #726 and verified expected behavior.
