### 2025-03-31 8fc163a
Made clever way to flatten multi-universe.
