### 2025-03-31 0e87752
Fixed iteration order for parsing fill matrix.
