### 2025-03-31 ae2fc7f
Added #726 to changelog.
