### 2025-03-31 14c7f4a
Switched to threads to preserve results.
