### 2025-03-31 #7951 @kyleniemeyer
@kellyrowland great! can you handle the last few steps of the fast-track publication (https://joss.readthedocs.io/en/latest/editing.html#ropensci-reviewed-or-pyopensci-reviewed-and-accepted-submissions) and then let me know when it's ready?

Basically, it's just the check the paper and share the software archive DOI here.
