### 2025-04-02 8aa4e35
Used @tjlaboss numpy wizardry.
