### 2025-04-02 b84739b
Rename & deprec Cell.lattice -> Cell.lattice_type
