### 2025-04-02 a46d909
Refactor Lattice -> LatticeType, and deprecate
