+++ montepy/__init__.py
@@ -18,7 +18,7 @@ from montepy.data_inputs.material import Material
 from montepy.data_inputs.transform import Transform
 from montepy.data_inputs.nuclide import Library, Nuclide
 from montepy.data_inputs.element import Element
-from montepy.data_inputs.lattice import Lattice
+from montepy.data_inputs.lattice import LatticeType
 from montepy.data_inputs.thermal_scattering import ThermalScatteringLaw
 from montepy.data_inputs.data_parser import parse_data
 
