### 2025-04-02 85cf1b3
Update changelog for #728
