### 2025-04-02 7ba5402
Rename Lattice -> LatticeType
