+++ montepy/data_inputs/lattice.py
@@ -3,18 +3,11 @@ from enum import Enum, unique
 from warnings import warn
 
 
-def _lattice_deprecation_warning():
-    warn(
-        message="lattice.Lattice is deprecated in favor of lattice.LatticeType",
-        category=DeprecationWarning,
-    )
-
-
 @unique
 class LatticeType(Enum):
     """Represents the options for the lattice ``LAT``."""
 
-    HEXAHEDRA = 1
+    HEXAHEDRAL = 1
     """Hexhedra are solids with six faces.
 
     One such solid is a rectangular prism.
@@ -23,15 +16,24 @@ class LatticeType(Enum):
     """Hexagonal prism are solids with eight faces."""
 
 
-class Lattice(Enum):
-    """Deprecated"""
+class __DeprecatedLattice:
+    """Helper for deprecated Enum behavior"""
 
     @property
     def HEXAHEDRA(self):
-        _lattice_deprecation_warning()
-        return LatticeType.HEXAHEDRA
+        warn(
+            message="Lattice.HEXAHEDRA is deprecated in favor of LatticeType.HEXAHEDRAL",
+            category=DeprecationWarning,
+        )
+        return LatticeType.HEXAHEDRAL
 
     @property
     def HEXAGONAL(self):
-        _lattice_deprecation_warning()
+        warn(
+            message="Lattice.HEXAGONAL is deprecated in favor of LatticeType.HEXAGONAL",
+            category=DeprecationWarning,
+        )
         return LatticeType.HEXAGONAL
+
+
+Lattice = __DeprecatedLattice()
