### 2025-04-02 8619d7a
Test for deprecation warnings for lattice renaming
