### 2025-04-02 9418858
Clean up the lattice test while we're at it.
