### 2025-04-02 085fffe
Initial implementation of mcnp_str
