### 2025-04-02 7ed2cce
Tested MCNP_str
