### 2025-04-03 cf47230
Fixed cell modifier format logic.
