### 2025-04-03 8848623
Suppress line expansion with mcnp_str.
