### 2025-04-03 c8ae10a
Updated cell test to avoid state errors.
