+++ montepy/data_inputs/cell_modifier.py
@@ -318,7 +318,8 @@ class CellModifierInput(DataInputAbstract):
                 mcnp_version = self._problem.mcnp_version
             else:
                 mcnp_version = montepy.MCNP_VERSION
-        with warnings.catch_warnings(action="ignore"):
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
             return "\n".join(
                 self.format_for_mcnp_input(mcnp_version, always_print=True)
             )
