### 2025-04-03 7740056
Tested mcnp_str wrapping and tried for 100% coverage.
