### 2025-04-03 8c0ca1f
Make sure MCNP_VERSION is what we expect.
