### 2025-04-03 2163c48
Don't mutate global variables!
