### 2025-04-04 b4d5062
prototyped using concurrent.futures.
