### 2025-04-06 ed329f9
Allowed None defaults.
