### 2025-04-06 248e241
Fixed issue with varargs.
