### 2025-04-06 306d52e
Updated functions for tests to be accurate for expected results.
