+++ tests/test_check_annote.py
@@ -10,6 +10,7 @@ import montepy._check_value as cv
 def str_annotations(a: str):
     pass
 
+
 @cv.check_arguments
 def str_annote_default(a: str = None):
     pass
@@ -38,6 +39,7 @@ def test_dummy_bad_type(val, func):
 @given(st.one_of(chars), funcs)
 def test_dummy_good_type(val, func):
     func(val)
+
 
 @given(st.one_of(chars, no), st.sampled_from([str_annote_default]))
 def test_none_default(val, func):
