### 2025-04-07 302d896
Tested additional value enforcement.
