### 2025-04-07 f6e8f04
Fixed bad function call test.
