+++ montepy/surface_collection.py
@@ -1,9 +1,9 @@
-# Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
+# Copyright 2024-2025, Battelle Energy Alliance, LLC All Rights Reserved.
 from __future__ import annotations
 import montepy
 from montepy.surfaces.surface import Surface
 from montepy.surfaces.surface_type import SurfaceType
-from montepy.numbered_object_collection import NumberedObjectCollection
+from montepy._abc.numbered_object_collection import NumberedObjectCollection
 
 
 def __create_surface_generator(surf_type):
