+++ montepy/data_inputs/material.py
@@ -13,7 +13,7 @@ from montepy.data_inputs.nuclide import Library, Nucleus, Nuclide
 from montepy.data_inputs.element import Element
 from montepy.input_parser import syntax_node
 from montepy.input_parser.material_parser import MaterialParser
-from montepy.numbered_mcnp_object import Numbered_MCNP_Object, InitInput
+from montepy._abc.numbered_mcnp_object import Numbered_MCNP_Object, InitInput
 from montepy.errors import *
 from montepy.utilities import *
 from montepy.particle import LibraryType
