### 2025-04-08 d849bbc
Fleshed out a prototype CoW
