### 2025-04-09 6d4f23e
Got metaclasses mostly working.
