### 2025-04-09 #7951 @kyleniemeyer
@MicahGale thanks. I'm now looking at the DOI... it looks like the DOI does not actually point to the Zenodo archive, but instead the OSTI / DOE CODE entry. However, that doesn't actually make the software archive available openly, which doesn't work for us. (I'm surprised pyOpenSci didn't flag that.)

I imagine you gave the DOI from DOE CODE when creating the Zenodo archive, but it should really be a separate one. In the Zenodo metadata, you could add the OSTI DOI as a supplemental link or reference.

Could you modify the Zenodo archive to get an actual Zenodo DOI that points to the software archive?
