### 2025-04-09 #7951 @kyleniemeyer
@MicahGale ok, just a few things to fix in the paper:
- There is an extraneous apostrophe after the OpenMC citation on page 3
- This isn't a big deal, but citations with Paul Romano as first author appear in two or three different ways ("P. Romano et al.", "P. K. Romano et al.", "Paul Romano et al.") because the `.bib` has the name formatted differently; BibTeX thinks these may be different names. Can you make sure the name is formatted the same throughout the `.bib` entries? I'd suggest just making all of them `P. K. Romano`.
- remove "[Journal Article]", "[Conference Paper]", "[Online multimedia]" (etc.) from the reference titles
