+++ inst/paper.bib
@@ -50,12 +50,11 @@
 }
 
 @article{Romano:2022,
-   author = {Romano, Paul K. and Stauff, Nicolas E. and Ooi, Zhiee Jhia and Miao, Yinbin and Lund, Amanda and Zou, Ling},
+   author = {Romano, P. K.  and Stauff, Nicolas E. and Ooi, Zhiee Jhia and Miao, Yinbin and Lund, Amanda and Zou, Ling},
    title = {WATTS: Workflow and template toolkit for simulation},
    journal = {Journal of Open Source Software},
    volume = {7},
    number = {79},
-   abstract = {Romano et al., (2022). WATTS: Workflow and template toolkit for simulation. Journal of Open Source Software, 7(79), 4735, https://doi.org/10.21105/joss.04735},
    ISSN = {2475-9066},
    DOI = {10.21105/joss.04735},
    url = {https://joss.theoj.org/papers/10.21105/joss.04735},
@@ -64,7 +63,7 @@
 }
 
 @article{Romano:2015,
-   author = {Romano, P. and Horelik, Nicholas E. and Herman, Bryan R. and Nelson, Adam G. and Forget, Benoit and Smith, Kord},
+   author = {Romano, P. K. and Horelik, Nicholas E. and Herman, Bryan R. and Nelson, Adam G. and Forget, Benoit and Smith, Kord},
    title = {OpenMC: A state-of-the-art Monte Carlo code for research and development},
    journal = {Annals of Nuclear Energy},
    volume = {82},
@@ -208,7 +207,7 @@
 @software{romano_mcnp_2024,
 	title = {{MCNP} Conversion tools for {OpenMC}},
 	url = {https://github.com/openmc-dev/openmc_mcnp_adapter},
-	author = {Romano, Paul and Li, Kaiwen and Shriwise, Patrick and Valderrama, Juan},
+	author = {Romano, P. K. and Li, Kaiwen and Shriwise, Patrick and Valderrama, Juan},
 	date = {2024},
 	keywords = {conversions, read input},
 }
