+++ inst/paper.bib
@@ -17,7 +17,6 @@
    title = {SLY (Sly Lex Yacc)},
    url = {https://sly.readthedocs.io/en/latest/},
    year = {2016},
-   type = {Online Multimedia}
 }
 
 @article{Kowal:2023,
@@ -31,7 +30,6 @@
    DOI = {10.1080/00295639.2022.2153617},
    url = {https://www.tandfonline.com/doi/full/10.1080/00295639.2022.2153617},
    year = {2023},
-   type = {Journal Article}
 }
 
 
@@ -46,7 +44,6 @@
    DOI = {10.1016/J.ENG.2017.04.016},
    url = {https://doi.org/10.1016/J.ENG.2017.04.016},
    year = {2017},
-   type = {Journal Article}
 }
 
 @article{Romano:2022,
@@ -59,7 +56,6 @@
    DOI = {10.21105/joss.04735},
    url = {https://joss.theoj.org/papers/10.21105/joss.04735},
    year = {2022},
-   type = {Journal Article}
 }
 
 @article{Romano:2015,
@@ -72,7 +68,6 @@
    DOI = {10.1016/j.anucene.2014.07.048},
    url = {https://doi.org/10.1016/j.anucene.2014.07.048},
    year = {2015},
-   type = {Journal Article}
 }
 
 @techreport{Bates:2022,
@@ -91,7 +86,7 @@
    title = {Advanced Test Reactor User Guide},
    institution = {Idaho National Laboratory},
    number = {INL/EXT-21-64328},
-   type = {User Guide},
+   type = {Report},
    month = {11 October},
    url = {https://www.osti.gov/biblio/1826354},
    DOI = {10.2172/1826354},
@@ -108,7 +103,6 @@
    pages = {985-987},
    month = {November 11-15},
    year = {2012},
-   type = {Conference Paper}
 }
 
 @software{rodionov_mckit_2024,
