### 2025-04-09 #7951 @MicahGale
@kyleniemeyer thanks for your attention to detail. The issue with the "[Journal article]"... comes from including `type={}` info a bibliography. I have removed most of those, but I did keep the `type={Report}`. I figure that context makes sense for providing a report number.

@kellyrowland thank you so much for all your help so far!
